Refactor: extract helper modules
