+++ AutoSafeguard.py
@@ -238,10 +238,14 @@ from review_toolbox import (
     ReviewDocumentDialog,
     VersionCompareDialog,
 )
-from dataclasses import dataclass, asdict, field
+from dataclasses import asdict
 from mechanisms import DiagnosticMechanism, MechanismLibrary, ANNEX_D_MECHANISMS
 import json
 import csv
+from drawing_helper import FTADrawingHelper, fta_drawing_helper
+from risk_assessment import DERIVED_MATURITY_TABLE, ASSURANCE_AGGREGATION_AND, AND_DECOMPOSITION_TABLE, OR_DECOMPOSITION_TABLE, boolify, ADRiskAssessmentHelper
+from models import MissionProfile, ReliabilityComponent, ReliabilityAnalysis, HazopEntry, HaraEntry, QUALIFICATIONS, COMPONENT_ATTR_TEMPLATES, RELIABILITY_MODELS, ASIL_LEVEL_OPTIONS, ASIL_ORDER, ASIL_TARGETS, ASIL_TABLE, calc_asil
+from toolboxes import ReliabilityWindow, FI2TCWindow, HazopWindow, HaraWindow, TC2FIWindow
 import copy
 import tkinter.font as tkFont
 from PIL import Image, ImageDraw, ImageFont, ImageTk
@@ -293,308 +297,8 @@ VALID_SUBTYPES = {
 # Global Unique ID Counter for Nodes
 ##########################################
 unique_node_id_counter = 1
-import math
 import uuid
 
-@dataclass
-class MissionProfile:
-    name: str
-    tau_on: float = 0.0
-    tau_off: float = 0.0
-    board_temp: float = 25.0
-    board_temp_min: float = 25.0
-    board_temp_max: float = 25.0
-    ambient_temp: float = 25.0
-    ambient_temp_min: float = 25.0
-    ambient_temp_max: float = 25.0
-    humidity: float = 50.0
-    duty_cycle: float = 1.0
-    notes: str = ""
-
-    @property
-    def temperature(self) -> float:
-        """Alias for backward compatibility (returns board temperature)."""
-        return self.board_temp
-
-    @temperature.setter
-    def temperature(self, value: float) -> None:
-        self.board_temp = value
-
-
-    @property
-    def tau(self) -> float:
-        """Return the total TAU for backward compatibility."""
-        return self.tau_on + self.tau_off
-
-@dataclass
-class ReliabilityComponent:
-    name: str
-    comp_type: str
-    quantity: int = 1
-    attributes: dict = field(default_factory=dict)
-    qualification: str = ""
-    fit: float = 0.0
-    is_passive: bool = False
-    sub_boms: list = field(default_factory=list)
-
-QUALIFICATIONS = [
-    "AEC-Q100",
-    "AEC-Q101",
-    "AEC-Q200",
-    "IECQ",
-    "MIL-STD-883",
-    "MIL-PRF-38534",
-    "MIL-PRF-38535",
-    "Space",
-    "None",
-]
-
-@dataclass
-class ReliabilityAnalysis:
-    """Store the results of a reliability calculation including the BOM."""
-
-    name: str
-    standard: str
-    profile: str
-    components: list
-    total_fit: float
-    spfm: float
-    lpfm: float
-    dc: float
-
-@dataclass
-class HazopEntry:
-    function: str
-    malfunction: str
-    mtype: str
-    scenario: str
-    conditions: str
-    hazard: str
-    safety: bool
-    rationale: str
-    covered: bool
-    covered_by: str
-    component: str = ""
-
-@dataclass
-class HaraEntry:
-    malfunction: str
-    severity: int
-    sev_rationale: str
-    controllability: int
-    cont_rationale: str
-    exposure: int
-    exp_rationale: str
-    asil: str
-    safety_goal: str
-
-COMPONENT_ATTR_TEMPLATES = {
-    "capacitor": {
-        "dielectric": ["ceramic", "electrolytic", "tantalum"],
-        "capacitance_uF": "",
-        "voltage_V": "",
-        "esr_ohm": "",
-        "tolerance_pct": "",
-    },
-    "resistor": {
-        "resistance_ohm": "",
-        "power_W": "",
-        "tolerance_pct": "",
-    },
-    "inductor": {
-        "inductance_H": "",
-        "current_A": "",
-        "saturation_A": "",
-    },
-    "diode": {
-        "type": ["standard", "zener", "schottky"],
-        "reverse_V": "",
-        "forward_current_A": "",
-    },
-    "transistor": {
-        "transistor_type": ["BJT", "MOSFET"],
-        "pins": "",
-        "voltage_V": "",
-        "current_A": "",
-    },
-    "ic": {
-        "type": ["digital", "analog", "mcu"],
-        "pins": "",
-        "transistors": "",
-    },
-    "connector": {"pins": ""},
-    "relay": {"cycles": "", "current_A": "", "voltage_V": ""},
-    "switch": {"cycles": "", "current_A": "", "voltage_V": ""},
-}
-
-RELIABILITY_MODELS = {
-    "IEC 62380": {
-        "capacitor": {
-            "text": "Base*(1+T/100)*Duty*(1+V/100)",
-            "formula": lambda a, mp: (0.02 if a.get("dielectric", "ceramic") == "ceramic" else 0.04)
-            * (1 + max(mp.board_temp_max, mp.ambient_temp_max) / 100.0)
-            * (1 + float(a.get("voltage_V", 0)) / 100.0)
-            * mp.duty_cycle,
-        },
-        "resistor": {
-            "text": "0.005*(1+T/100)*Duty*(1+P/10)",
-            "formula": lambda a, mp: 0.005
-            * (1 + max(mp.board_temp_max, mp.ambient_temp_max) / 100.0)
-            * (1 + float(a.get("power_W", 0)) / 10.0)
-            * mp.duty_cycle,
-        },
-        "inductor": {
-            "text": "0.004*(1+T/100)*Duty*(1+I/10)",
-            "formula": lambda a, mp: 0.004
-            * (1 + max(mp.board_temp_max, mp.ambient_temp_max) / 100.0)
-            * (1 + float(a.get("current_A", 0)) / 10.0)
-            * mp.duty_cycle,
-        },
-        "diode": {
-            "text": "0.008*(1+T/100)*Duty*(1+VR/100)",
-            "formula": lambda a, mp: 0.008
-            * (1 + max(mp.board_temp_max, mp.ambient_temp_max) / 100.0)
-            * (1 + float(a.get("reverse_V", 0)) / 100.0)
-            * mp.duty_cycle,
-        },
-        "transistor": {
-            "text": "Base*(1+T/100)*Duty*(1+I/10) (base=0.01 BJT, 0.012 MOSFET)",
-            "formula": lambda a, mp: (0.01 if a.get("transistor_type", "BJT") == "BJT" else 0.012)
-            * (1 + max(mp.board_temp_max, mp.ambient_temp_max) / 100.0)
-            * (1 + float(a.get("current_A", 0)) / 10.0)
-            * mp.duty_cycle,
-        },
-        "ic": {
-            "text": "Base*(1+T/100)*Duty*(1+pins/1000)*(1+trans/1e6) (base=0.04 analog, 0.03 digital, 0.05 MCU)",
-            "formula": lambda a, mp: (0.04 if a.get("type", "digital") == "analog" else 0.05 if a.get("type") == "mcu" else 0.03)
-            * (1 + max(mp.board_temp_max, mp.ambient_temp_max) / 100.0)
-            * mp.duty_cycle
-            * (1 + float(a.get("pins", 0)) / 1000.0)
-            * (1 + float(a.get("transistors", 0)) / 1_000_000.0),
-        },
-        "connector": {
-            "text": "0.002*(1+T/100)*Duty*(1+pins/100)",
-            "formula": lambda a, mp: 0.002
-            * (1 + max(mp.board_temp_max, mp.ambient_temp_max) / 100.0)
-            * (1 + float(a.get("pins", 0)) / 100.0)
-            * mp.duty_cycle,
-        },
-        "relay": {
-            "text": "0.03*(1+T/100)*Duty*(cycles/1e6)*(1+I/10)",
-            "formula": lambda a, mp: 0.03
-            * (1 + max(mp.board_temp_max, mp.ambient_temp_max) / 100.0)
-            * mp.duty_cycle
-            * (float(a.get("cycles", 1e6)) / 1e6)
-            * (1 + float(a.get("current_A", 0)) / 10.0),
-        },
-        "switch": {
-            "text": "0.02*(1+T/100)*Duty*(cycles/1e6)*(1+I/10)",
-            "formula": lambda a, mp: 0.02
-            * (1 + max(mp.board_temp_max, mp.ambient_temp_max) / 100.0)
-            * mp.duty_cycle
-            * (float(a.get("cycles", 1e6)) / 1e6)
-            * (1 + float(a.get("current_A", 0)) / 10.0),
-        },
-    },
-    "SN 29500": {
-        "capacitor": {
-            "text": "0.03*(1+T/100)*Duty*(1+V/100)",
-            "formula": lambda a, mp: 0.03
-            * (1 + max(mp.board_temp_max, mp.ambient_temp_max) / 100.0)
-            * (1 + float(a.get("voltage_V", 0)) / 100.0)
-            * mp.duty_cycle,
-        },
-        "resistor": {
-            "text": "0.006*(1+T/100)*Duty*(1+P/10)",
-            "formula": lambda a, mp: 0.006
-            * (1 + max(mp.board_temp_max, mp.ambient_temp_max) / 100.0)
-            * (1 + float(a.get("power_W", 0)) / 10.0)
-            * mp.duty_cycle,
-        },
-        "inductor": {
-            "text": "0.005*(1+T/100)*Duty*(1+I/10)",
-            "formula": lambda a, mp: 0.005
-            * (1 + max(mp.board_temp_max, mp.ambient_temp_max) / 100.0)
-            * (1 + float(a.get("current_A", 0)) / 10.0)
-            * mp.duty_cycle,
-        },
-        "diode": {
-            "text": "0.009*(1+T/100)*Duty*(1+VR/100)",
-            "formula": lambda a, mp: 0.009
-            * (1 + max(mp.board_temp_max, mp.ambient_temp_max) / 100.0)
-            * (1 + float(a.get("reverse_V", 0)) / 100.0)
-            * mp.duty_cycle,
-        },
-        "transistor": {
-            "text": "Base*(1+T/100)*Duty*(1+I/10) (base=0.012 BJT, 0.014 MOSFET)",
-            "formula": lambda a, mp: (0.012 if a.get("transistor_type", "BJT") == "BJT" else 0.014)
-            * (1 + max(mp.board_temp_max, mp.ambient_temp_max) / 100.0)
-            * (1 + float(a.get("current_A", 0)) / 10.0)
-            * mp.duty_cycle,
-        },
-        "ic": {
-            "text": "Base*(1+T/100)*Duty*(1+pins/1000)*(1+trans/1e6) (base=0.05 analog, 0.04 digital, 0.06 MCU)",
-            "formula": lambda a, mp: (0.05 if a.get("type", "digital") == "analog" else 0.06 if a.get("type") == "mcu" else 0.04)
-            * (1 + max(mp.board_temp_max, mp.ambient_temp_max) / 100.0)
-            * mp.duty_cycle
-            * (1 + float(a.get("pins", 0)) / 1000.0)
-            * (1 + float(a.get("transistors", 0)) / 1_000_000.0),
-        },
-        "connector": {
-            "text": "0.003*(1+T/100)*Duty*(1+pins/100)",
-            "formula": lambda a, mp: 0.003
-            * (1 + max(mp.board_temp_max, mp.ambient_temp_max) / 100.0)
-            * (1 + float(a.get("pins", 0)) / 100.0)
-            * mp.duty_cycle,
-        },
-        "relay": {
-            "text": "0.035*(1+T/100)*Duty*(cycles/1e6)*(1+I/10)",
-            "formula": lambda a, mp: 0.035
-            * (1 + max(mp.board_temp_max, mp.ambient_temp_max) / 100.0)
-            * mp.duty_cycle
-            * (float(a.get("cycles", 1e6)) / 1e6)
-            * (1 + float(a.get("current_A", 0)) / 10.0),
-        },
-        "switch": {
-            "text": "0.025*(1+T/100)*Duty*(cycles/1e6)*(1+I/10)",
-            "formula": lambda a, mp: 0.025
-            * (1 + max(mp.board_temp_max, mp.ambient_temp_max) / 100.0)
-            * mp.duty_cycle
-            * (float(a.get("cycles", 1e6)) / 1e6)
-            * (1 + float(a.get("current_A", 0)) / 10.0),
-        },
-    },
-}
-
-global_requirements = {}
-# ASIL level options including decomposition levels
-ASIL_LEVEL_OPTIONS = [
-    "QM", "QM(A)", "QM(B)", "QM(C)", "QM(D)",
-    "A", "A(B)", "B", "B(C)", "C", "C(D)", "D"
-]
-
-ASIL_ORDER = {"QM":0, "A":1, "B":2, "C":3, "D":4}
-ASIL_TARGETS = {
-    "D": {"spfm":0.99, "lpfm":0.90, "dc":0.99},
-    "C": {"spfm":0.97, "lpfm":0.90, "dc":0.97},
-    "B": {"spfm":0.90, "lpfm":0.60, "dc":0.90},
-    "A": {"spfm":0.0, "lpfm":0.0, "dc":0.0},
-    "QM": {"spfm":0.0, "lpfm":0.0, "dc":0.0},
-}
-
-# Simplified ISO 26262 risk graph for ASIL determination
-ASIL_TABLE = {
-    (3, 1, 4): "D", (3, 1, 3): "D", (3, 1, 2): "C", (3, 1, 1): "B",
-    (3, 2, 4): "C", (3, 2, 3): "C", (3, 2, 2): "B", (3, 2, 1): "A",
-    (3, 3, 4): "B", (3, 3, 3): "B", (3, 3, 2): "A", (3, 3, 1): "QM",
-    (2, 1, 4): "C", (2, 1, 3): "C", (2, 1, 2): "B", (2, 1, 1): "A",
-    (2, 2, 4): "B", (2, 2, 3): "B", (2, 2, 2): "A", (2, 2, 1): "QM",
-    (2, 3, 4): "A", (2, 3, 3): "A", (2, 3, 2): "QM", (2, 3, 1): "QM",
-}
-
-def calc_asil(sev: int, cont: int, expo: int) -> str:
-    """Return ASIL based on severity, controllability and exposure."""
-    return ASIL_TABLE.get((sev, cont, expo), "QM")
 dynamic_recommendations = {
     1: {
         "Testing Requirements": (
@@ -957,959 +661,6 @@ dynamic_recommendations = {
     }
 }
 
-# Derived Maturity Table: (avg_confidence, avg_robustness) → maturity level
-DERIVED_MATURITY_TABLE = {
-    (1, 1): 1, (1, 2): 1, (1, 3): 1, (1, 4): 2, (1, 5): 2,
-    (2, 2): 2, (2, 3): 2, (2, 4): 3, (2, 5): 3,
-    (3, 3): 3, (3, 4): 3, (3, 5): 4,
-    (4, 4): 4, (4, 5): 4,
-    (5, 5): 5,
-}
-
-ASSURANCE_AGGREGATION_AND = {
-    (1,1): 3,
-    (1,2): 4, (2,2): 4,
-    (1,3): 4, (2,3): 4, (3,3): 5,
-    (1,4): 5, (2,4): 5, (3,4): 5, (4,4): 5,
-    (1,5): 5, (2,5): 5, (3,5): 5, (4,5): 5, (5,5): 5
-}
-
-AND_DECOMPOSITION_TABLE = {
-    3: [(1, 1)],
-    4: [(1, 2), (2, 2), (1, 3), (2, 3)],
-    5: [(1, 4), (2, 4), (3, 4), (4, 4),
-        (1, 5), (2, 5), (3, 5), (4, 5), (5, 5)]
-}
-
-OR_DECOMPOSITION_TABLE = {
-    1: [(5, 5)],
-    2: [(4, 4)],
-    3: [(3, 3)],
-    4: [(2, 2)],
-    5: [(1, 1)]
-}
-    
-def boolify(value, default):
-    if isinstance(value, str):
-        return value.lower() == "true"
-    return bool(value) if value is not None else default
-        
-class ADRiskAssessmentHelper:
-    """
-    Helper class for risk assessment computations.
-    It encapsulates methods to:
-      - Generate unique node IDs.
-      - Update the unique ID counter based on a list of top events.
-      - Round a value to the nearest half.
-      - Discretize a continuous value into a level from 1 to 5.
-      - Combine input values based on gate type.
-      - Recursively calculate assurance (or maturity/rigor) values.
-    """
-    def __init__(self):
-        self.unique_node_id_counter = 1
-
-    def aggregate_clone_requirements(self, clone_node):
-        """
-        If the given node is a clone, then:
-          - For each child in the original node, collect its safety requirements.
-          - Gather safety goals from the clone's own parents and the original node's parents.
-          - Link (i.e. add) these safety goals to each of the collected requirements.
-        
-        Returns a dictionary keyed by requirement key (its "id" if available, else its text)
-        with each value containing:
-             "req": the requirement dictionary,
-             "linked_sgs": a set of safety goal strings.
-        """
-        # Only process if node is a clone
-        if not clone_node.is_primary_instance and hasattr(clone_node, "original") and clone_node.original:
-            aggregated = {}
-
-            # 1. Collect requirements from each child of the original node.
-            # (Assume that the safety requirements live on the base events.)
-            children_reqs = []
-            for child in clone_node.original.children:
-                # You might want to further traverse children if needed; here we assume direct children.
-                if hasattr(child, "safety_requirements") and child.safety_requirements:
-                    children_reqs.extend(child.safety_requirements)
-                else:
-                    # Optionally, if child has its own children, traverse downward.
-                    def collect_reqs(n):
-                        reqs = []
-                        if hasattr(n, "safety_requirements") and n.safety_requirements:
-                            reqs.extend(n.safety_requirements)
-                        for c in n.children:
-                            reqs.extend(collect_reqs(c))
-                        return reqs
-                    children_reqs.extend(collect_reqs(child))
-            
-            # 2. Gather safety goals from the clone's immediate parents.
-            clone_parent_goals = set()
-            for parent in clone_node.parents:
-                if parent.safety_goal_description and parent.safety_goal_description.strip():
-                    clone_parent_goals.add(f"- {parent.safety_goal_description.strip()}")
-                else:
-                    clone_parent_goals.add(f"- {parent.name}")
-            
-            # 3. Also gather safety goals from the original node's immediate parents.
-            original_parent_goals = set()
-            for parent in clone_node.original.parents:
-                if parent.safety_goal_description and parent.safety_goal_description.strip():
-                    original_parent_goals.add(f"- {parent.safety_goal_description.strip()}")
-                else:
-                    original_parent_goals.add(f"- {parent.name}")
-            
-            # Union both sets.
-            safety_goals = clone_parent_goals.union(original_parent_goals)
-            print(f"DEBUG: For clone node {clone_node.unique_id}, clone_parent_goals={clone_parent_goals}, original_parent_goals={original_parent_goals}")
-
-            # 4. For each collected requirement, add the safety goals.
-            for req in children_reqs:
-                key = req.get("id") if req.get("id") else req.get("text", "Unnamed Requirement")
-                if key not in aggregated:
-                    aggregated[key] = {
-                        "req": req,
-                        "linked_sgs": set()
-                    }
-                aggregated[key]["linked_sgs"].update(safety_goals)
-                print(f"DEBUG: Linking safety goals {safety_goals} to requirement {key} from original child")
-            return aggregated
-        else:
-            # If not a clone, return an empty dict (or handle as needed)
-            return {}
-
-    def fix_clone_references(self, root_nodes):
-        # First pass: collect all primary nodes from every top event.
-        primary_by_id = {}
-        def collect_primary(node):
-            if node.is_primary_instance:
-                primary_by_id[node.unique_id] = node
-                print(f"[DEBUG] Added primary node: id={node.unique_id}, name='{node.user_name}'")
-            for child in node.children:
-                collect_primary(child)
-        for root in root_nodes:
-            collect_primary(root)
-        
-        # Second pass: update all clones using the complete dictionary.
-        def fix(node):
-            if not node.is_primary_instance:
-                orig_id = getattr(node, "_original_id", node.unique_id)
-                print(f"[DEBUG] Fixing clone: id={node.unique_id}, _original_id={orig_id}")
-                if orig_id in primary_by_id:
-                    node.original = primary_by_id[orig_id]
-                    print(f"[DEBUG] Clone {node.unique_id} now references primary node {node.original.unique_id}")
-                else:
-                    node.original = node
-                    print(f"[DEBUG] No matching primary for clone {node.unique_id} with _original_id={orig_id}; using self")
-            else:
-                node.original = node
-                print(f"[DEBUG] Primary node {node.unique_id} set to reference itself")
-            for child in node.children:
-                fix(child)
-        for root in root_nodes:
-            fix(root)
-
-    def get_next_unique_id(self):
-        uid = self.unique_node_id_counter
-        self.unique_node_id_counter += 1
-        return uid
-
-    def update_unique_id_counter_for_top_events(self, top_events):
-        def traverse(node):
-            ids = [node.unique_id]
-            for child in node.children:
-                ids.extend(traverse(child))
-            return ids
-        all_ids = []
-        for event in top_events:
-            all_ids.extend(traverse(event))
-        self.unique_node_id_counter = max(all_ids) + 1
-
-    def round_to_half(self, val):
-        try:
-            val = float(val)
-        except Exception as e:
-            print(f"Error converting {val} to float: {e}")
-            val = 0.0
-        return round(val * 2) / 2
-
-    def discretize_level(self, val):
-        #r = self.round_to_half(val)
-        r = val
-        if r < 1.5:
-            return 1
-        elif r < 2.5:
-            return 2
-        elif r < 3.5:
-            return 3
-        elif r < 4.5:
-            return 4
-        else:
-            return 5
-
-    def scale_severity(self, sev):
-        """Map severity 1-3 to a 1-5 scale."""
-        try:
-            sev = float(sev)
-        except Exception:
-            sev = 3
-        sev = max(1.0, min(3.0, sev))
-        return (sev - 1) * 2 + 1
-
-    def scale_controllability(self, cont):
-        """Map controllability 1-3 to a 1-5 scale."""
-        try:
-            cont = float(cont)
-        except Exception:
-            cont = 3
-        cont = max(1.0, min(3.0, cont))
-        return (cont - 1) * 2 + 1
-
-    def combine_values(self, values, gate_type):
-        if not values:
-            return 1.0
-        if gate_type.upper() == "AND":
-            prod = 1.0
-            for v in values:
-                prod *= (1 - v/5)
-            return (1 - prod) * 5
-        else:
-            return sum(values) / len(values)
-
-    def combine_rigor_or(self,values):
-        # Using the reliability (complement-product) formula.
-        prod = 1.0
-        for v in values:
-            prod *= (1 - v/5)
-        return round((1 - prod) * 5, 2)
-        
-    def combine_rigor_and(self,values):
-        return sum(values) / len(values)            
-            
-    def combine_generic_values(self, values, gate_type):
-        if not values:
-            return None
-        gate_type = gate_type.upper()
-        if gate_type == "AND":
-            prod = 1.0
-            for v in values:
-                prod *= (1 - round(v/5, 2))
-            return round((1 - prod) * 5, 2)
-        else:
-            return round(sum(values) / len(values), 2)
-
-    def is_effectively_confidence(self,node):
-        """
-        Returns True if the node is either:
-          - A base event with node_type "CONFIDENCE LEVEL", or
-          - A gate (or similar) whose children are all effectively confidence.
-        """
-        if node.node_type.upper() == "CONFIDENCE LEVEL":
-            return True
-        if node.children and node.node_type.upper() in ["GATE", "RIGOR LEVEL", "TOP EVENT"]:
-            return all(self.is_effectively_confidence(child) for child in node.children)
-        return False
-
-    def is_effectively_robustness(self,node):
-        """
-        Returns True if the node is either:
-          - A base event with node_type "ROBUSTNESS SCORE", or
-          - A gate (or similar) whose children are all effectively robustness.
-        """
-        if node.node_type.upper() == "ROBUSTNESS SCORE":
-            return True
-        if node.children and node.node_type.upper() in ["GATE", "RIGOR LEVEL", "TOP EVENT"]:
-            return all(self.is_effectively_robustness(child) for child in node.children)
-        return False
-
-    def aggregate_assurance_and(self,child_levels):
-        """
-        Combine a list of children’s Prototype Assurance Levels (PAL) for an AND gate,
-        using pairwise lookups in ASSURANCE_AGGREGATION_AND.
-        """
-        if not child_levels:
-            return 1
-        current = child_levels[0]
-        for next_val in child_levels[1:]:
-            pair = tuple(sorted((current, next_val)))
-            # If not found in the dict, fallback to max(...) or something:
-            current = ASSURANCE_AGGREGATION_AND.get(pair, max(pair))
-        return current
-
-    def aggregate_assurance_or(self,child_levels):
-        if not child_levels:
-            return 1
-        avg = sum(child_levels) // len(child_levels)
-        return max(1, min(5, avg))
-
-    def derive_assurance_from_base(self,conf_values, rob_values):
-        """
-        Given lists of confidence and robustness integers (each 1..5),
-        compute a single 'inverted' Prototype Assurance Level (PAL) from 1..5,
-        where low confidence/robustness inputs produce a high assurance value.
-        """
-        if not conf_values or not rob_values:
-            return 1  # fallback
-        # Compute the integer average for each
-        avg_conf = round(sum(conf_values) / len(conf_values))
-        avg_rob  = round(sum(rob_values) / len(rob_values))
-        # Hard-coded 5×5 assurance matrix:
-        assurance_matrix = [
-          [5, 4, 4, 3, 3],  # Confidence = 1
-          [4, 4, 3, 3, 2],  # Confidence = 2
-          [4, 3, 3, 2, 2],  # Confidence = 3
-          [3, 3, 2, 2, 1],  # Confidence = 4
-          [3, 2, 2, 1, 1]   # Confidence = 5
-        ]
-        # Adjust indices (1 maps to index 0, etc.)
-        c_idx = max(1, min(5, avg_conf)) - 1
-        r_idx = max(1, min(5, avg_rob)) - 1
-        return assurance_matrix[c_idx][r_idx]
-
-    def get_highest_parent_severity_for_node(self, node, all_top_events):
-        """
-        Return the highest severity found among all ancestors of all instances
-        (primary or clone) of 'node' across every top event in 'all_top_events'.
-        If no ancestor has a valid severity, return 3 by default.
-        """
-        # 1) Identify the primary ID for the node
-        primary_id = node.unique_id if node.is_primary_instance else node.original.unique_id
-
-        # 2) Collect all instances (primary or clones) with that primary ID from all top events.
-        instances = []
-        def collect_instances(root):
-            def walk(n):
-                if n.is_primary_instance and n.unique_id == primary_id:
-                    instances.append(n)
-                elif (not n.is_primary_instance and n.original and 
-                      n.original.unique_id == primary_id):
-                    instances.append(n)
-                for c in n.children:
-                    walk(c)
-            walk(root)
-        for te in all_top_events:
-            collect_instances(te)
-
-        # 3) Traverse upward (using DFS) from each instance to find the maximum severity.
-        visited = set()
-        max_sev = 0
-        def dfs_up(n):
-            nonlocal max_sev
-            if n in visited:
-                return
-            visited.add(n)
-            if n.severity is not None:
-                try:
-                    s = int(n.severity)
-                    if s > max_sev:
-                        max_sev = s
-                except:
-                    pass
-            for p in n.parents:
-                dfs_up(p)
-            # For clones, also check the original's parents.
-            if (not n.is_primary_instance) and n.original and (n.original != n):
-                for p2 in n.original.parents:
-                    dfs_up(p2)
-        for inst in instances:
-            dfs_up(inst)
-        return max_sev if max_sev > 0 else 3
-
-    def aggregate_assurance_or_adjusted(self, child_levels):
-        """
-        For an OR gate, compute the average of the child levels and then invert the result using a 6 - average rule.
-        For example, if the average child level is 4 (strong), then 6 - 4 = 2, meaning the overall assurance requirement is 2.
-        Ensure the final value is between 1 and 5.
-        """
-        if not child_levels:
-            return 1
-        avg = sum(child_levels) / len(child_levels)
-        inverted = 6 - avg
-        return max(1, min(5, round(inverted)))
-
-    def calculate_assurance_recursive(self, node, all_top_events, visited=None):
-        if visited is None:
-            visited = set()
-        if node.unique_id in visited:
-            return node.quant_value if node.quant_value is not None else 1
-        visited.add(node.unique_id)
-        t = node.node_type.upper()
-
-        # --- Base Events ---
-        if t == "CONFIDENCE LEVEL":
-            cval = max(1, min(5, int(node.quant_value if node.quant_value is not None else 1)))
-            node.quant_value = cval
-            node.display_label = f"Confidence [{cval}]"
-            node.detailed_equation = f"Base Confidence => {cval}"
-            return cval
-        if t == "ROBUSTNESS SCORE":
-            rval = max(1, min(5, int(node.quant_value if node.quant_value is not None else 1)))
-            node.quant_value = rval
-            node.display_label = f"Robustness [{rval}]"
-            node.detailed_equation = f"Base Robustness => {rval}"
-            return rval
-
-        if not node.children:
-            fallback = max(1, min(5, int(node.quant_value if node.quant_value is not None else 1)))
-            node.quant_value = fallback
-            node.display_label = f"Node [{fallback}]"
-            node.detailed_equation = f"No children => fallback value {fallback}"
-            return fallback
-
-        # Process all children recursively.
-        for child in node.children:
-            self.calculate_assurance_recursive(child, all_top_events, visited)
-
-        # --- Separate children into base events and composite children ---
-        base_values = []
-        composite_values = []
-        for child in node.children:
-            ctype = child.node_type.upper()
-            if ctype in ["CONFIDENCE LEVEL", "ROBUSTNESS SCORE"]:
-                # Use the already computed quant_value (which is now the assurance value)
-                base_values.append(max(1, min(5, int(child.quant_value))))
-            else:
-                composite_values.append(max(1, min(5, int(child.quant_value))))
-
-        # For the base events, if present, compute the assurance using our inversion matrix.
-        if base_values:
-            # When only one type is present, we use the same list for both inputs.
-            base_assurance = self.derive_assurance_from_base(base_values, base_values)
-        else:
-            base_assurance = None
-
-        # For composite children, aggregate their assurance values using the appropriate gate rule.
-        gate = (node.gate_type or "AND").upper()
-        if composite_values:
-            if gate == "AND":
-                composite_assurance = self.aggregate_assurance_and(composite_values)
-            elif gate == "OR":
-                composite_assurance = self.aggregate_assurance_or_adjusted(composite_values)
-            else:
-                composite_assurance = None
-        else:
-            composite_assurance = None
-
-        # Combine base assurance and composite assurance.
-        if base_assurance is not None and composite_assurance is not None:
-            combined = (base_assurance + composite_assurance) // 2
-        elif base_assurance is not None:
-            combined = base_assurance
-        elif composite_assurance is not None:
-            combined = composite_assurance
-        else:
-            combined = 1
-
-        level_map = {1: "extra low", 2: "low", 3: "moderate", 4: "high", 5: "high+"}
-
-        if node.node_type.upper() == "TOP EVENT":
-            try:
-                s_raw = float(node.severity)
-            except (TypeError, ValueError):
-                s_raw = 3
-            s = self.scale_severity(s_raw)
-            
-            try:
-                c_raw = float(node.controllability)
-            except (TypeError, ValueError):
-                c_raw = 3
-            c = self.scale_controllability(c_raw)
-
-            final = round((combined + s + c) / 3)
-            final = max(1, min(5, final))
-            node.quant_value = final
-            node.display_label = f"Prototype Assurance Level (PAL) [{level_map[final]}]"
-            node.detailed_equation = (
-                f"Base Assurance from children = {base_assurance if base_assurance is not None else 'N/A'}\n"
-                f"Composite Assurance from gates = {composite_assurance if composite_assurance is not None else 'N/A'}\n"
-                f"Combined (average) = {combined}\n"
-                f"Node Severity (TOP EVENT) = {s_raw} (scaled: {s})\n"
-                f"Node Controllability = {c_raw} (scaled: {c})\n"
-                f"Final Assurance = (({combined} + {s} + {c}) /3) = {final}"
-            )
-            return final
-        else:
-            node.quant_value = combined
-            node.display_label = f"Prototype Assurance Level (PAL) [{level_map[combined]}]"
-            node.detailed_equation = (
-                f"Base Assurance from children = {base_assurance if base_assurance is not None else 'N/A'}\n"
-                f"Composite Assurance from gates = {composite_assurance if composite_assurance is not None else 'N/A'}\n"
-                f"Combined Children Assurance (average) = {combined}\n"
-            )
-            return combined
-
-    def calculate_probability_recursive(self, node, visited=None):
-        """Return the probability of failure for ``node``.
-
-        The method traverses the fault tree bottom-up, combining child
-        probabilities according to the node's ``gate_type``.  For an AND gate
-        the probabilities are multiplied, while an OR gate uses the
-        ``1 - \u220f(1 - p)`` rule.  Basic events simply return their assigned
-        probability.
-        """
-        if visited is None:
-            visited = set()
-
-        # Avoid infinite recursion but allow the same node to be evaluated
-        # along different branches when it appears more than once.
-        if node.unique_id in visited:
-            return node.probability if node.probability is not None else 0.0
-
-        visited.add(node.unique_id)
-        t = node.node_type.upper()
-        if t == "BASIC EVENT":
-            prob = float(node.failure_prob)
-            node.probability = prob
-            node.display_label = f"P={prob:.2e}"
-            return prob
-
-        if not node.children:
-            prob = float(getattr(node, "failure_prob", 0.0))
-            node.probability = prob
-            node.display_label = f"P={prob:.2e}"
-            return prob
-
-        # Use a fresh visited set for each child to ensure probabilities
-        # propagate correctly even when subtrees are shared between gates.
-        child_probs = [self.calculate_probability_recursive(c, visited.copy()) for c in node.children]
-
-        gate = (node.gate_type or "AND").upper()
-        if gate == "AND":
-            prob = 1.0
-            for p in child_probs:
-                prob *= p
-        else:
-            prod = 1.0
-            for p in child_probs:
-                prod *= (1 - p)
-            prob = 1 - prod
-        node.probability = prob
-        node.display_label = f"P={prob:.2e}"
-        return prob
-
-
-class FTADrawingHelper:
-    """
-    A helper class that provides drawing functions for fault tree diagrams.
-    These methods can be used to draw shapes (gates, events, connectors, etc.)
-    onto a tkinter Canvas.
-    """
-    def __init__(self):
-        pass
-
-    def get_text_size(self, text, font_obj):
-        """Return the (width, height) in pixels needed to render the text with the given font."""
-        lines = text.split("\n")
-        max_width = max(font_obj.measure(line) for line in lines)
-        height = font_obj.metrics("linespace") * len(lines)
-        return max_width, height
-
-    def draw_page_clone_shape(self, canvas, x, y, scale=40.0,
-                              top_text="Desc:\n\nRationale:", bottom_text="Node",
-                              fill="lightgray", outline_color="dimgray",
-                              line_width=1, font_obj=None):
-        # First, draw the main triangle using the existing triangle routine.
-        self.draw_triangle_shape(canvas, x, y, scale=scale,
-                                 top_text=top_text, bottom_text=bottom_text,
-                                 fill=fill, outline_color=outline_color,
-                                 line_width=line_width, font_obj=font_obj)
-        # Determine a baseline for the bottom of the triangle.
-        # (You may need to adjust this value to match your triangle's dimensions.)
-        bottom_y = y + scale * 0.75  
-        # Draw two horizontal lines at the bottom
-        line_offset1 = scale * 0.05
-        line_offset2 = scale * 0.1
-        canvas.create_line(x - scale/2, bottom_y - line_offset1,
-                           x + scale/2, bottom_y - line_offset1,
-                           fill=outline_color, width=line_width)
-        canvas.create_line(x - scale/2, bottom_y - line_offset2,
-                           x + scale/2, bottom_y - line_offset2,
-                           fill=outline_color, width=line_width)
-        # Draw a small triangle on the right side as a clone indicator.
-        tri_side = scale * 0.5
-        tri_height = (math.sqrt(3) / 2) * tri_side
-        att_x = x + scale  # position to the right of the main triangle
-        att_y = y - tri_height / 2 - tri_height# adjust vertical position as needed
-        v1 = (att_x, att_y)
-        v2 = (att_x + tri_side, att_y)
-        v3 = (att_x + tri_side/2, att_y - tri_height)
-        canvas.create_polygon(v1, v2, v3, fill="lightblue", outline=outline_color,
-                              width=line_width)
-
-    def draw_shared_marker(self, canvas, x, y, zoom):
-        """Draw a small shared marker at the given canvas coordinates."""
-        size = 10 * zoom
-        v1 = (x, y)
-        v2 = (x - size, y)
-        v3 = (x, y - size)
-        canvas.create_polygon([v1, v2, v3], fill="black", outline="black")
-
-    def draw_90_connection(self, canvas, parent_pt, child_pt, outline_color="dimgray", line_width=1, fixed_length=40):
-        """Draw a 90° connection line from a parent point to a child point."""
-        fixed_y = parent_pt[1] + fixed_length
-        canvas.create_line(parent_pt[0], parent_pt[1], parent_pt[0], fixed_y,
-                           fill=outline_color, width=line_width)
-        canvas.create_line(parent_pt[0], fixed_y, child_pt[0], fixed_y,
-                           fill=outline_color, width=line_width)
-        canvas.create_line(child_pt[0], fixed_y, child_pt[0], child_pt[1],
-                           fill=outline_color, width=line_width)
-
-    def compute_rotated_and_gate_vertices(self, scale):
-        """Compute vertices for a rotated AND gate shape scaled by 'scale'."""
-        vertices = [(0, 0), (0, 2), (1, 2)]
-        num_points = 50
-        for i in range(num_points + 1):
-            theta = math.pi / 2 - math.pi * i / num_points
-            vertices.append((1 + math.cos(theta), 1 + math.sin(theta)))
-        vertices.append((0, 0))
-        def rotate_point(pt):
-            x, y = pt
-            return (2 - y, x)
-        rotated = [rotate_point(pt) for pt in vertices]
-        translated = [(vx + 2, vy + 1) for (vx, vy) in rotated]
-        scaled = [(vx * scale, vy * scale) for (vx, vy) in translated]
-        return scaled
-
-    def draw_rotated_and_gate_shape(self, canvas, x, y, scale=40.0,
-                                      top_text="Desc:\n\nRationale:",
-                                      bottom_text="Event",
-                                      fill="lightgray", outline_color="dimgray",
-                                      line_width=1, font_obj=None):
-        """Draw a rotated AND gate shape with top and bottom text labels."""
-        if font_obj is None:
-            font_obj = tkFont.Font(family="Arial", size=10)
-        raw_verts = self.compute_rotated_and_gate_vertices(scale)
-        flipped = [(vx, -vy) for (vx, vy) in raw_verts]
-        xs = [v[0] for v in flipped]
-        ys = [v[1] for v in flipped]
-        cx, cy = (sum(xs) / len(xs), sum(ys) / len(ys))
-        final_points = [(vx - cx + x, vy - cy + y) for (vx, vy) in flipped]
-        canvas.create_polygon(final_points, fill=fill, outline=outline_color,
-                                width=line_width, smooth=False)
-
-        # Draw the top label box
-        t_width, t_height = self.get_text_size(top_text, font_obj)
-        padding = 6
-        top_box_width = t_width + 2 * padding
-        top_box_height = t_height + 2 * padding
-        top_y = min(pt[1] for pt in final_points) - top_box_height - 5
-        top_box_x = x - top_box_width / 2
-        canvas.create_rectangle(top_box_x, top_y,
-                                top_box_x + top_box_width,
-                                top_y + top_box_height,
-                                fill="lightblue", outline=outline_color,
-                                width=line_width)
-        canvas.create_text(top_box_x + top_box_width / 2,
-                           top_y + top_box_height / 2,
-                           text=top_text,
-                           font=font_obj,
-                           anchor="center",
-                           width=top_box_width)
-
-        # Draw the bottom label box
-        b_width, b_height = self.get_text_size(bottom_text, font_obj)
-        bottom_box_width = b_width + 2 * padding
-        bottom_box_height = b_height + 2 * padding
-        shape_lowest_y = max(pt[1] for pt in final_points)
-        bottom_y = shape_lowest_y - (2 * bottom_box_height)
-        bottom_box_x = x - bottom_box_width / 2
-        canvas.create_rectangle(bottom_box_x, bottom_y,
-                                bottom_box_x + bottom_box_width,
-                                bottom_y + bottom_box_height,
-                                fill="lightblue", outline=outline_color,
-                                width=line_width)
-        canvas.create_text(bottom_box_x + bottom_box_width / 2,
-                           bottom_y + bottom_box_height / 2,
-                           text=bottom_text,
-                           font=font_obj,
-                           anchor="center",
-                           width=bottom_box_width)
-
-    def draw_rotated_or_gate_shape(self, canvas, x, y, scale=40.0,
-                                     top_text="Desc:\n\nRationale:",
-                                     bottom_text="Event",
-                                     fill="lightgray", outline_color="dimgray",
-                                     line_width=1, font_obj=None):
-        """Draw a rotated OR gate shape with text labels."""
-        if font_obj is None:
-            font_obj = tkFont.Font(family="Arial", size=10)
-        def cubic_bezier(P0, P1, P2, P3, t):
-            return ((1 - t) ** 3 * P0[0] + 3 * (1 - t) ** 2 * t * P1[0] +
-                    3 * (1 - t) * t ** 2 * P2[0] + t ** 3 * P3[0],
-                    (1 - t) ** 3 * P0[1] + 3 * (1 - t) ** 2 * t * P1[1] +
-                    3 * (1 - t) * t ** 2 * P2[1] + t ** 3 * P3[1])
-        num_points = 30
-        t_values = [i / num_points for i in range(num_points + 1)]
-        seg1 = [cubic_bezier((0, 0), (0.6, 0), (0.6, 2), (0, 2), t) for t in t_values]
-        seg2 = [cubic_bezier((0, 2), (1, 2), (2, 1.6), (2, 1), t) for t in t_values]
-        seg3 = [cubic_bezier((2, 1), (2, 0.4), (1, 0), (0, 0), t) for t in t_values]
-        points = seg1[:-1] + seg2[:-1] + seg3
-        rotated = [(2 - p[1], p[0]) for p in points]
-        translated = [(pt[0] + 2, pt[1] + 1) for pt in rotated]
-        scaled = [(sx * scale, sy * scale) for (sx, sy) in translated]
-        flipped = [(vx, -vy) for (vx, vy) in scaled]
-        xs = [p[0] for p in flipped]
-        ys = [p[1] for p in flipped]
-        cx, cy = (sum(xs) / len(xs), sum(ys) / len(ys))
-        final_points = [(vx - cx + x, vy - cy + y) for (vx, vy) in flipped]
-        canvas.create_polygon(final_points, fill=fill, outline=outline_color,
-                                width=line_width, smooth=True)
-
-        # Draw the top label box
-        padding = 6
-        t_width, t_height = self.get_text_size(top_text, font_obj)
-        top_box_width = t_width + 2 * padding
-        top_box_height = t_height + 2 * padding
-        top_y = min(pt[1] for pt in final_points) - top_box_height - 5
-        top_box_x = x - top_box_width / 2
-        canvas.create_rectangle(top_box_x, top_y,
-                                top_box_x + top_box_width,
-                                top_y + top_box_height,
-                                fill="lightblue", outline=outline_color, width=line_width)
-        canvas.create_text(top_box_x + top_box_width / 2,
-                           top_y + top_box_height / 2,
-                           text=top_text, font=font_obj, anchor="center",
-                           width=top_box_width)
-
-        # Draw the bottom label box
-        b_width, b_height = self.get_text_size(bottom_text, font_obj)
-        bottom_box_width = b_width + 2 * padding
-        bottom_box_height = b_height + 2 * padding
-        shape_lowest_y = max(pt[1] for pt in final_points)
-        bottom_y = shape_lowest_y - (2 * bottom_box_height)
-        bottom_box_x = x - bottom_box_width / 2
-        canvas.create_rectangle(bottom_box_x, bottom_y,
-                                bottom_box_x + bottom_box_width,
-                                bottom_y + bottom_box_height,
-                                fill="lightblue", outline=outline_color,
-                                width=line_width)
-        canvas.create_text(bottom_box_x + bottom_box_width / 2,
-                           bottom_y + bottom_box_height / 2,
-                           text=bottom_text, font=font_obj,
-                           anchor="center", width=bottom_box_width)
-
-    def draw_rotated_and_gate_clone_shape(self, canvas, x, y, scale=40.0,
-                                            top_text="Desc:\n\nRationale:", bottom_text="Node",
-                                            fill="lightgray", outline_color="dimgray",
-                                            line_width=1, font_obj=None):
-        """Draw a rotated AND gate shape with additional clone details."""
-        self.draw_rotated_and_gate_shape(canvas, x, y, scale=scale,
-                                         top_text=top_text, bottom_text=bottom_text,
-                                         fill=fill, outline_color=outline_color,
-                                         line_width=line_width, font_obj=font_obj)
-        bottom_y = y + scale * 1.5
-        line_offset1 = scale * 0.05
-        line_offset2 = scale * 0.1
-        canvas.create_line(x - scale/2, bottom_y - line_offset1,
-                           x + scale/2, bottom_y - line_offset1,
-                           fill=outline_color, width=line_width)
-        canvas.create_line(x - scale/2, bottom_y - line_offset2,
-                           x + scale/2, bottom_y - line_offset2,
-                           fill=outline_color, width=line_width)
-        tri_side = scale * 0.5
-        tri_height = (math.sqrt(3) / 2) * tri_side
-        att_x = x + scale
-        att_y = y - tri_height / 2
-        v1 = (att_x, att_y)
-        v2 = (att_x + tri_side, att_y)
-        v3 = (att_x + tri_side / 2, att_y - tri_height)
-        canvas.create_polygon(v1, v2, v3, fill="lightblue", outline=outline_color,
-                              width=line_width)
-        final_line_offset = scale * 0.15
-        canvas.create_line(x - scale/2, bottom_y + final_line_offset,
-                           x + scale/2, bottom_y + final_line_offset,
-                           fill=outline_color, width=line_width)
-
-    def draw_rotated_or_gate_clone_shape(self, canvas, x, y, scale=40.0,
-                                           top_text="Desc:\n\nRationale:", bottom_text="Node",
-                                           fill="lightgray", outline_color="dimgray",
-                                           line_width=1, font_obj=None):
-        """Draw a rotated OR gate shape with additional clone details."""
-        self.draw_rotated_or_gate_shape(canvas, x, y, scale=scale,
-                                        top_text=top_text, bottom_text=bottom_text,
-                                        fill=fill, outline_color=outline_color,
-                                        line_width=line_width, font_obj=font_obj)
-        bottom_y = y + scale * 1.5
-        line_offset1 = scale * 0.05
-        line_offset2 = scale * 0.1
-        canvas.create_line(x - scale/2, bottom_y - line_offset1,
-                           x + scale/2, bottom_y - line_offset1,
-                           fill=outline_color, width=line_width)
-        canvas.create_line(x - scale/2, bottom_y - line_offset2,
-                           x + scale/2, bottom_y - line_offset2,
-                           fill=outline_color, width=line_width)
-        tri_side = scale * 0.5
-        tri_height = (math.sqrt(3) / 2) * tri_side
-        att_x = x + scale
-        att_y = y - tri_height / 2
-        v1 = (att_x, att_y)
-        v2 = (att_x + tri_side, att_y)
-        v3 = (att_x + tri_side / 2, att_y - tri_height)
-        canvas.create_polygon(v1, v2, v3, fill="lightblue",
-                              outline=outline_color, width=line_width)
-        final_line_offset = scale * 0.15
-        canvas.create_line(x - scale/2, bottom_y + final_line_offset,
-                           x + scale/2, bottom_y + final_line_offset,
-                           fill=outline_color, width=line_width)
-
-    def draw_triangle_shape(self, canvas, x, y, scale=40.0,
-                              top_text="Desc:\n\nRationale:",
-                              bottom_text="Event",
-                              fill="lightgray", outline_color="dimgray",
-                              line_width=1, font_obj=None):
-        if font_obj is None:
-            font_obj = tkFont.Font(family="Arial", size=10)
-        effective_scale = scale * 2  
-        h = effective_scale * math.sqrt(3) / 2
-        v1 = (0, -2 * h / 3)
-        v2 = (-effective_scale / 2, h / 3)
-        v3 = (effective_scale / 2, h / 3)
-        vertices = [(x + v1[0], y + v1[1]),
-                    (x + v2[0], y + v2[1]),
-                    (x + v3[0], y + v3[1])]
-        canvas.create_polygon(vertices, fill=fill, outline=outline_color, width=line_width)
-        
-        t_width, t_height = self.get_text_size(top_text, font_obj)
-        padding = 6
-        top_box_width = t_width + 2 * padding
-        top_box_height = t_height + 2 * padding
-        top_box_x = x - top_box_width / 2
-        top_box_y = min(v[1] for v in vertices) - top_box_height
-        canvas.create_rectangle(top_box_x, top_box_y,
-                                top_box_x + top_box_width,
-                                top_box_y + top_box_height,
-                                fill="lightblue", outline=outline_color, width=line_width)
-        canvas.create_text(top_box_x + top_box_width / 2,
-                           top_box_y + top_box_height / 2,
-                           text=top_text,
-                           font=font_obj, anchor="center", width=top_box_width)
-        
-        b_width, b_height = self.get_text_size(bottom_text, font_obj)
-        bottom_box_width = b_width + 2 * padding
-        bottom_box_height = b_height + 2 * padding
-        bottom_box_x = x - bottom_box_width / 2
-        bottom_box_y = max(v[1] for v in vertices) + padding - 2 * bottom_box_height
-        canvas.create_rectangle(bottom_box_x, bottom_box_y,
-                                bottom_box_x + bottom_box_width,
-                                bottom_box_y + bottom_box_height,
-                                fill="lightblue", outline=outline_color, width=line_width)
-        canvas.create_text(bottom_box_x + bottom_box_width / 2,
-                           bottom_box_y + bottom_box_height / 2,
-                           text=bottom_text,
-                           font=font_obj, anchor="center", width=bottom_box_width)
-                           
-    def draw_circle_event_shape(self, canvas, x, y, radius,
-                                top_text="",
-                                bottom_text="",
-                                fill="lightyellow",
-                                outline_color="dimgray",
-                                line_width=1,
-                                font_obj=None,
-                                base_event=False):
-        """Draw a circular event shape with optional text labels."""
-        if font_obj is None:
-            font_obj = tkFont.Font(family="Arial", size=10)
-        left = x - radius
-        top = y - radius
-        right = x + radius
-        bottom = y + radius
-        canvas.create_oval(left, top, right, bottom, fill=fill,
-                           outline=outline_color, width=line_width)
-        t_width, t_height = self.get_text_size(top_text, font_obj)
-        padding = 6
-        top_box_width = t_width + 2 * padding
-        top_box_height = t_height + 2 * padding
-        top_box_x = x - top_box_width / 2
-        top_box_y = top - top_box_height
-        canvas.create_rectangle(top_box_x, top_box_y,
-                                top_box_x + top_box_width,
-                                top_box_y + top_box_height,
-                                fill="lightblue", outline=outline_color,
-                                width=line_width)
-        canvas.create_text(top_box_x + top_box_width / 2,
-                           top_box_y + top_box_height / 2,
-                           text=top_text,
-                           font=font_obj, anchor="center",
-                           width=top_box_width)
-        b_width, b_height = self.get_text_size(bottom_text, font_obj)
-        bottom_box_width = b_width + 2 * padding
-        bottom_box_height = b_height + 2 * padding
-        bottom_box_x = x - bottom_box_width / 2
-        bottom_box_y = bottom - 2 * bottom_box_height
-        canvas.create_rectangle(bottom_box_x, bottom_box_y,
-                                bottom_box_x + bottom_box_width,
-                                bottom_box_y + bottom_box_height,
-                                fill="lightblue", outline=outline_color,
-                                width=line_width)
-        canvas.create_text(bottom_box_x + bottom_box_width / 2,
-                           bottom_box_y + bottom_box_height / 2,
-                           text=bottom_text,
-                           font=font_obj, anchor="center",
-                           width=bottom_box_width)
-                           
-    def draw_triangle_clone_shape(self, canvas, x, y, scale=40.0,
-                                  top_text="Desc:\n\nRationale:", bottom_text="Node",
-                                  fill="lightgray", outline_color="dimgray",
-                                  line_width=1, font_obj=None):
-        """
-        Draws the same triangle as draw_triangle_shape but then adds two horizontal lines
-        at the bottom and a small triangle on the right side as clone indicators.
-        The small triangle is now positioned so that its top vertex aligns with the top of
-        the big triangle.
-        """
-        if font_obj is None:
-            font_obj = tkFont.Font(family="Arial", size=10)
-        # Draw the base triangle.
-        self.draw_triangle_shape(canvas, x, y, scale=scale,
-                                 top_text=top_text, bottom_text=bottom_text,
-                                 fill=fill, outline_color=outline_color,
-                                 line_width=line_width, font_obj=font_obj)
-        # Compute the vertices of the big triangle.
-        effective_scale = scale * 2  
-        h = effective_scale * math.sqrt(3) / 2
-        v1 = (0, -2 * h / 3)
-        v2 = (-effective_scale / 2, h / 3)
-        v3 = (effective_scale / 2, h / 3)
-        vertices = [(x + v1[0], y + v1[1]),
-                    (x + v2[0], y + v2[1]),
-                    (x + v3[0], y + v3[1])]
-        # Compute the bottom and top y-values of the big triangle.
-        bottom_y = max(v[1] for v in vertices) + scale * 0.2
-        top_y = min(v[1] for v in vertices)  # top edge of the big triangle
-        half_width = effective_scale / 2  # equals 'scale'
-        
-        # Draw two horizontal lines at the bottom (unchanged).
-        line_offset1 = scale * 0.05
-        line_offset2 = scale * 0.1
-        canvas.create_line(x - half_width, bottom_y - line_offset1,
-                           x + half_width, bottom_y - line_offset1,
-                           fill=outline_color, width=line_width)
-        canvas.create_line(x - half_width, bottom_y - line_offset2,
-                           x + half_width, bottom_y - line_offset2,
-                           fill=outline_color, width=line_width)
-        
-        # Draw the small clone indicator triangle.
-        tri_side = scale * 0.5
-        tri_height = (math.sqrt(3) / 2) * tri_side
-        att_x = x + half_width
-        # Instead of basing its vertical position on bottom_y, we now align it with top_y.
-        # We want the top vertex of the small triangle (which is at att_y - tri_height)
-        # to equal top_y. Thus, set att_y - tri_height = top_y, so:
-        att_y = top_y + tri_height
-        v1_small = (att_x, att_y)
-        v2_small = (att_x + tri_side, att_y)
-        v3_small = (att_x + tri_side/2, att_y - tri_height)
-        canvas.create_polygon(v1_small, v2_small, v3_small,
-                              fill="lightblue", outline=outline_color,
-                              width=line_width)
-        
-        # Draw the final horizontal line below the bottom.
-        final_line_offset = scale * 0.15
-        canvas.create_line(x - half_width, bottom_y + final_line_offset,
-                           x + half_width, bottom_y + final_line_offset,
-                           fill=outline_color, width=line_width)
-                           
-# Create a single FTADrawingHelper object that can be used by other classes
-fta_drawing_helper = FTADrawingHelper()
 AD_RiskAssessment_Helper = ADRiskAssessmentHelper()
 
 ##########################################
@@ -10406,7 +9157,7 @@ class FaultTreeApp:
         if hasattr(self, "_rel_window") and self._rel_window.winfo_exists():
             self._rel_window.lift()
             return
-        self._rel_window = self.ReliabilityWindow(self)
+        self._rel_window = ReliabilityWindow(self)
 
     def open_fmeda_window(self):
         self.show_fmeda_list()
@@ -10415,1327 +9166,26 @@ class FaultTreeApp:
         if hasattr(self, "_hazop_window") and self._hazop_window.winfo_exists():
             self._hazop_window.lift()
             return
-        self._hazop_window = self.HazopWindow(self)
+        self._hazop_window = HazopWindow(self)
 
     def open_hara_window(self):
         if hasattr(self, "_hara_window") and self._hara_window.winfo_exists():
             self._hara_window.lift()
             return
-        self._hara_window = self.HaraWindow(self)
+        self._hara_window = HaraWindow(self)
 
     def open_fi2tc_window(self):
         if hasattr(self, "_fi2tc_window") and self._fi2tc_window.winfo_exists():
             self._fi2tc_window.lift()
             return
-        self._fi2tc_window = self.FI2TCWindow(self)
+        self._fi2tc_window = FI2TCWindow(self)
 
     def open_tc2fi_window(self):
         if hasattr(self, "_tc2fi_window") and self._tc2fi_window.winfo_exists():
             self._tc2fi_window.lift()
             return
-        self._tc2fi_window = self.TC2FIWindow(self)
+        self._tc2fi_window = TC2FIWindow(self)
         
-    class ReliabilityWindow(tk.Toplevel):
-        def __init__(self, app):
-            super().__init__(app.root)
-            self.app = app
-            self.title("Reliability Analysis")
-            self.geometry("600x400")
-            self.components = []
-
-            ttk.Label(self, text="Standard:").pack(anchor="w")
-            self.standard_var = tk.StringVar(value="IEC 62380")
-            ttk.Combobox(
-                self,
-                textvariable=self.standard_var,
-                values=["IEC 62380", "SN 29500"],
-                state="readonly",
-            ).pack(anchor="w")
-
-            ttk.Label(self, text="Mission Profile:").pack(anchor="w")
-            self.profile_var = tk.StringVar()
-            self.profile_combo = ttk.Combobox(
-                self,
-                textvariable=self.profile_var,
-                values=[mp.name for mp in app.mission_profiles],
-                state="readonly",
-            )
-            self.profile_combo.pack(anchor="w", fill="x")
-
-            self.tree = ttk.Treeview(
-                self,
-                columns=("name", "type", "qty", "fit", "qualification"),
-                show="headings",
-            )
-            for col in ("name", "type", "qty", "fit", "qualification"):
-                heading = "Qualification" if col == "qualification" else col.capitalize()
-                self.tree.heading(col, text=heading)
-                self.tree.column(col, width=120 if col == "qualification" else 100)
-            self.tree.pack(fill=tk.BOTH, expand=True)
-            self.tree.bind("<<TreeviewSelect>>", self.show_formula)
-
-            btn_frame = ttk.Frame(self)
-            btn_frame.pack(fill=tk.X)
-            ttk.Button(btn_frame, text="Load CSV", command=self.load_csv).pack(
-                side=tk.LEFT, padx=2, pady=2
-            )
-            ttk.Button(btn_frame, text="Add Component", command=self.add_component).pack(
-                side=tk.LEFT, padx=2, pady=2
-            )
-            ttk.Button(btn_frame, text="Add Circuit", command=self.add_circuit).pack(
-                side=tk.LEFT, padx=2, pady=2
-            )
-            ttk.Button(btn_frame, text="Configure Component", command=self.configure_component).pack(
-                side=tk.LEFT, padx=2, pady=2
-            )
-            ttk.Button(btn_frame, text="Calculate FIT", command=self.calculate_fit).pack(
-                side=tk.LEFT, padx=2, pady=2
-            )
-            ttk.Button(btn_frame, text="Save Analysis", command=self.save_analysis).pack(
-                side=tk.LEFT, padx=2, pady=2
-            )
-            ttk.Button(btn_frame, text="Load Analysis", command=self.load_analysis).pack(
-                side=tk.LEFT, padx=2, pady=2
-            )
-            self.formula_label = ttk.Label(self, text="")
-            self.formula_label.pack(anchor="w", padx=5, pady=5)
-
-        def add_component(self):
-            dialog = tk.Toplevel(self)
-            dialog.title("New Component")
-            ttk.Label(dialog, text="Name").grid(row=0, column=0, padx=5, pady=5, sticky="e")
-            name_var = tk.StringVar()
-            ttk.Entry(dialog, textvariable=name_var).grid(row=0, column=1, padx=5, pady=5)
-            ttk.Label(dialog, text="Type").grid(row=1, column=0, padx=5, pady=5, sticky="e")
-            type_var = tk.StringVar(value="capacitor")
-            ttk.Combobox(dialog, textvariable=type_var, values=list(COMPONENT_ATTR_TEMPLATES.keys()), state="readonly").grid(row=1, column=1, padx=5, pady=5)
-            ttk.Label(dialog, text="Quantity").grid(row=2, column=0, padx=5, pady=5, sticky="e")
-            qty_var = tk.IntVar(value=1)
-            ttk.Entry(dialog, textvariable=qty_var).grid(row=2, column=1, padx=5, pady=5)
-            ttk.Label(dialog, text="Qualification").grid(row=3, column=0, padx=5, pady=5, sticky="e")
-            qual_var = tk.StringVar(value="None")
-            ttk.Combobox(dialog, textvariable=qual_var, values=QUALIFICATIONS, state="readonly").grid(row=3, column=1, padx=5, pady=5)
-            passive_var = tk.BooleanVar(value=False)
-            ttk.Checkbutton(dialog, text="Passive", variable=passive_var).grid(row=4, column=0, columnspan=2, pady=5)
-
-            def ok():
-                comp = ReliabilityComponent(
-                    name_var.get(),
-                    type_var.get(),
-                    qty_var.get(),
-                    {},
-                    qual_var.get(),
-                    is_passive=passive_var.get(),
-                )
-                template = COMPONENT_ATTR_TEMPLATES.get(comp.comp_type, {})
-                for k, v in template.items():
-                    comp.attributes[k] = v[0] if isinstance(v, list) else v
-                self.components.append(comp)
-                self.refresh_tree()
-                dialog.destroy()
-
-            ttk.Button(dialog, text="Add", command=ok).grid(row=5, column=0, columnspan=2, pady=5)
-            dialog.grab_set()
-            dialog.wait_window()
-
-        def add_circuit(self):
-            dlg = tk.Toplevel(self)
-            dlg.title("New Circuit")
-            ttk.Label(dlg, text="Name").grid(row=0, column=0, padx=5, pady=5, sticky="e")
-            name_var = tk.StringVar()
-            ttk.Entry(dlg, textvariable=name_var).grid(row=0, column=1, padx=5, pady=5)
-            ttk.Label(dlg, text="BOMs").grid(row=1, column=0, padx=5, pady=5, sticky="ne")
-            lb = tk.Listbox(dlg, selectmode=tk.MULTIPLE, height=6)
-            for ra in self.app.reliability_analyses:
-                lb.insert(tk.END, ra.name)
-            lb.grid(row=1, column=1, padx=5, pady=5)
-            qty_var = tk.IntVar(value=1)
-            ttk.Label(dlg, text="Quantity").grid(row=2, column=0, padx=5, pady=5, sticky="e")
-            ttk.Entry(dlg, textvariable=qty_var).grid(row=2, column=1, padx=5, pady=5)
-            ttk.Label(dlg, text="Qualification").grid(row=3, column=0, padx=5, pady=5, sticky="e")
-            qual_var = tk.StringVar(value="None")
-            ttk.Combobox(dlg, textvariable=qual_var, values=QUALIFICATIONS, state="readonly").grid(row=3, column=1, padx=5, pady=5)
-            passive_var = tk.BooleanVar(value=False)
-            ttk.Checkbutton(dlg, text="Passive", variable=passive_var).grid(row=4, column=0, columnspan=2, pady=5)
-
-            def ok():
-                bom_idxs = lb.curselection()
-                boms = [self.app.reliability_analyses[i].components for i in bom_idxs]
-                comp = ReliabilityComponent(
-                    name_var.get(),
-                    "circuit",
-                    qty_var.get(),
-                    {},
-                    qual_var.get(),
-                    is_passive=passive_var.get(),
-                )
-                comp.sub_boms = copy.deepcopy(boms)
-                comp.fit = 0.0
-                self.components.append(comp)
-                self.refresh_tree()
-                dlg.destroy()
-
-            ttk.Button(dlg, text="Add", command=ok).grid(row=5, column=0, columnspan=2, pady=5)
-            dlg.grab_set()
-            dlg.wait_window()
-
-        def show_formula(self, event=None):
-            sel = self.tree.focus()
-            if not sel:
-                self.formula_label.config(text="")
-                return
-            idx = self.tree.index(sel)
-            if idx >= len(self.components):
-                return
-            comp = self.components[idx]
-            std = self.standard_var.get()
-            info = RELIABILITY_MODELS.get(std, {}).get(comp.comp_type)
-            if info:
-                self.formula_label.config(text=f"Formula: {info['text']}")
-            else:
-                self.formula_label.config(text="Formula: N/A")
-
-        def refresh_tree(self):
-            self.tree.delete(*self.tree.get_children())
-            for comp in self.components:
-                self.tree.insert(
-                    "",
-                    "end",
-                    values=(
-                        comp.name,
-                        comp.comp_type,
-                        comp.quantity,
-                        f"{comp.fit:.2f}",
-                        comp.qualification,
-                    ),
-                )
-            self.profile_combo.config(values=[mp.name for mp in self.app.mission_profiles])
-
-        def load_csv(self):
-            path = filedialog.askopenfilename(filetypes=[("CSV", "*.csv")])
-            if not path:
-                return
-            self.components.clear()
-            with open(path, newline="") as f:
-                reader = csv.DictReader(f)
-                fields = reader.fieldnames or []
-                mapping = self.ask_mapping(fields)
-                if not mapping:
-                    return
-                for row in reader:
-                    try:
-                        name = row.get(mapping["name"], "")
-                        ctype = row.get(mapping["type"], "")
-                        qty = int(row.get(mapping["qty"], 1) or 1)
-                        qual = row.get(mapping.get("qualification"), "") if mapping.get("qualification") else ""
-                        comp = ReliabilityComponent(name, ctype, qty, {}, qual)
-                        template = COMPONENT_ATTR_TEMPLATES.get(ctype, {})
-                        for k, v in template.items():
-                            comp.attributes[k] = v[0] if isinstance(v, list) else v
-                        # store any extra columns as attributes
-                        for key, val in row.items():
-                            if key not in mapping.values():
-                                comp.attributes[key] = val
-                        self.components.append(comp)
-                    except Exception:
-                        continue
-            self.refresh_tree()
-
-        def ask_mapping(self, fields):
-            if not fields:
-                return None
-            win = tk.Toplevel(self)
-            win.title("Map Columns")
-            vars = {}
-            targets = ["name", "type", "qty", "qualification"]
-            for i, tgt in enumerate(targets):
-                ttk.Label(win, text=tgt.capitalize()).grid(row=i, column=0, padx=5, pady=5, sticky="e")
-                var = tk.StringVar()
-                cb = ttk.Combobox(win, textvariable=var, values=fields, state="readonly")
-                if i < len(fields):
-                    var.set(fields[i])
-                cb.grid(row=i, column=1, padx=5, pady=5)
-                vars[tgt] = var
-
-            result = {}
-
-            def ok():
-                for k, v in vars.items():
-                    result[k] = v.get()
-                win.destroy()
-
-            def cancel():
-                result.clear()
-                win.destroy()
-
-            ttk.Button(win, text="OK", command=ok).grid(row=len(targets), column=0, pady=5)
-            ttk.Button(win, text="Cancel", command=cancel).grid(row=len(targets), column=1, pady=5)
-            win.grab_set()
-            win.wait_window()
-            if not result:
-                return None
-            return result
-
-        def configure_component(self):
-            sel = self.tree.focus()
-            if not sel:
-                messagebox.showwarning("Configure", "Select a component")
-                return
-            idx = self.tree.index(sel)
-            comp = self.components[idx]
-
-            template = COMPONENT_ATTR_TEMPLATES.get(comp.comp_type, {})
-            for k, v in template.items():
-                comp.attributes.setdefault(k, v[0] if isinstance(v, list) else v)
-
-            class ParamDialog(simpledialog.Dialog):
-                def body(self, master):
-                    self.vars = {}
-                    row = 0
-                    ttk.Label(master, text="Quantity").grid(row=row, column=0, padx=5, pady=5, sticky="e")
-                    qty_var = tk.IntVar(value=comp.quantity)
-                    ttk.Entry(master, textvariable=qty_var).grid(row=row, column=1, padx=5, pady=5)
-                    self.vars["__qty__"] = qty_var
-                    row += 1
-                    ttk.Label(master, text="Qualification").grid(row=row, column=0, padx=5, pady=5, sticky="e")
-                    qual_var = tk.StringVar(value=comp.qualification)
-                    ttk.Combobox(master, textvariable=qual_var, values=QUALIFICATIONS, state="readonly").grid(row=row, column=1, padx=5, pady=5)
-                    self.vars["__qual__"] = qual_var
-                    row += 1
-                    for k, v in comp.attributes.items():
-                        ttk.Label(master, text=k).grid(row=row, column=0, padx=5, pady=5, sticky="e")
-                        if isinstance(template.get(k), list):
-                            var = tk.StringVar(value=str(v))
-                            ttk.Combobox(master, textvariable=var, values=template[k], state="readonly").grid(row=row, column=1, padx=5, pady=5)
-                        else:
-                            var = tk.StringVar(value=str(v))
-                            ttk.Entry(master, textvariable=var).grid(row=row, column=1, padx=5, pady=5)
-                        self.vars[k] = var
-                        row += 1
-
-                def apply(self):
-                    comp.quantity = int(self.vars["__qty__"].get())
-                    comp.qualification = self.vars["__qual__"].get()
-                    for k, v in self.vars.items():
-                        if k.startswith("__"):
-                            continue
-                        comp.attributes[k] = v.get()
-
-            ParamDialog(self)
-            self.refresh_tree()
-
-        def calculate_fit(self):
-            prof_name = self.profile_var.get()
-            mp = next((m for m in self.app.mission_profiles if m.name == prof_name), None)
-            if mp is None:
-                messagebox.showwarning("FIT", "Select a mission profile")
-                return
-            std = self.standard_var.get()
-            total = 0.0
-            for comp in self.components:
-                if comp.sub_boms:
-                    sub_total = 0.0
-                    for bom in comp.sub_boms:
-                        for sub in bom:
-                            info = RELIABILITY_MODELS.get(std, {}).get(sub.comp_type)
-                            if info:
-                                sub.fit = info["formula"](sub.attributes, mp) * mp.tau
-                            else:
-                                sub.fit = 0.0
-                            sub_total += sub.fit * sub.quantity
-                    comp.fit = sub_total
-                else:
-                    info = RELIABILITY_MODELS.get(std, {}).get(comp.comp_type)
-                    if info:
-                        comp.fit = info["formula"](comp.attributes, mp) * mp.tau
-                    else:
-                        comp.fit = 0.0
-                total += comp.fit * comp.quantity
-
-            comp_fit = {c.name: c.fit * c.quantity for c in self.components}
-            spf = 0.0
-            lpf = 0.0
-            total_modes = 0.0
-            for be in self.app.fmea_entries:
-                comp_name = (
-                    be.parents[0].user_name if be.parents else getattr(be, "fmea_component", "")
-                )
-                fit = comp_fit.get(comp_name, 0.0)
-                frac = be.fmeda_fault_fraction
-                if frac > 1.0:
-                    frac /= 100.0
-                fit_mode = fit * frac
-                total_modes += fit_mode
-                if be.fmeda_fault_type == "permanent":
-                    spf += fit_mode * (1 - be.fmeda_diag_cov)
-                else:
-                    lpf += fit_mode * (1 - be.fmeda_diag_cov)
-            dc = 1 - (spf + lpf) / total_modes if total_modes else 0.0
-            self.app.reliability_components = list(self.components)
-            self.app.reliability_total_fit = total_modes
-            self.app.spfm = spf
-            self.app.lpfm = lpf
-            self.app.reliability_dc = dc
-            self.refresh_tree()
-            self.formula_label.config(
-                text=f"Total FIT: {total_modes:.2f}  DC: {dc:.2f}  SPFM: {spf:.2f}  LPFM: {lpf:.2f}"
-            )
-
-        def save_analysis(self):
-            if not self.components:
-                messagebox.showwarning("Save", "No components defined")
-                return
-            name = simpledialog.askstring("Save Analysis", "Enter analysis name:")
-            if not name:
-                return
-            ra = ReliabilityAnalysis(
-                name,
-                self.standard_var.get(),
-                self.profile_var.get(),
-                copy.deepcopy(self.components),
-                self.app.reliability_total_fit,
-                self.app.spfm,
-                self.app.lpfm,
-                self.app.reliability_dc,
-            )
-            self.app.reliability_analyses.append(ra)
-            messagebox.showinfo("Save", "Analysis saved")
-
-
-    class FI2TCWindow(tk.Toplevel):
-        COLS = [
-            "id","system_function","allocation","interfaces","insufficiency",
-            "scene","scenario","driver_behavior","occurrence","vehicle_effect",
-            "severity","design_measures","verification","measure_effectiveness",
-            "triggering_condition","worst_case","tc_effect","mitigation","acceptance"
-        ]
-        def __init__(self, app):
-            super().__init__(app.root)
-            self.app = app
-            self.title("FI2TC Analysis")
-            self.tree = ttk.Treeview(self, columns=self.COLS, show="headings")
-            for c in self.COLS:
-                self.tree.heading(c, text=c.replace("_"," ").title())
-                self.tree.column(c, width=120)
-            self.tree.pack(fill=tk.BOTH, expand=True)
-            btn = ttk.Frame(self)
-            btn.pack()
-            ttk.Button(btn, text="Add", command=self.add_row).pack(side=tk.LEFT, padx=2, pady=2)
-            ttk.Button(btn, text="Edit", command=self.edit_row).pack(side=tk.LEFT, padx=2, pady=2)
-            ttk.Button(btn, text="Delete", command=self.del_row).pack(side=tk.LEFT, padx=2, pady=2)
-            ttk.Button(btn, text="Export CSV", command=self.export_csv).pack(side=tk.LEFT, padx=2, pady=2)
-            self.refresh()
-
-        def refresh(self):
-            self.tree.delete(*self.tree.get_children())
-            for row in self.app.fi2tc_entries:
-                self.tree.insert("", "end", values=[row.get(k, "") for k in self.COLS])
-
-        class RowDialog(simpledialog.Dialog):
-            def __init__(self, parent, app, data=None):
-                self.app = app
-                self.parent_win = parent
-                default = {k: "" for k in parent.COLS}
-                self.data = data or default
-                super().__init__(parent, title="Edit Row")
-            def body(self, master):
-                fi_names = [n.user_name or f"FI {n.unique_id}" for n in self.app.get_all_functional_insufficiencies()]
-                tc_names = [n.user_name or f"TC {n.unique_id}" for n in self.app.get_all_triggering_conditions()]
-                self.widgets = {}
-                r = 0
-                for col in self.parent_win.COLS:
-                    ttk.Label(master, text=col.replace("_"," ").title()).grid(row=r, column=0, sticky="e")
-                    if col == "triggering_condition":
-                        var = tk.StringVar(value=self.data.get(col, ""))
-                        cb = ttk.Combobox(master, textvariable=var, values=tc_names, state="readonly")
-                        cb.grid(row=r, column=1)
-                        self.widgets[col] = var
-                    elif col == "insufficiency":
-                        var = tk.StringVar(value=self.data.get(col, ""))
-                        cb = ttk.Combobox(master, textvariable=var, values=fi_names, state="readonly")
-                        cb.grid(row=r, column=1)
-                        self.widgets[col] = var
-                    else:
-                        ent = tk.Entry(master)
-                        ent.insert(0, self.data.get(col, ""))
-                        ent.grid(row=r, column=1)
-                        self.widgets[col] = ent
-                    r += 1
-            def apply(self):
-                for col, widget in self.widgets.items():
-                    if isinstance(widget, tk.Entry):
-                        self.data[col] = widget.get()
-                    else:
-                        self.data[col] = widget.get()
-
-        def add_row(self):
-            dlg = self.RowDialog(self, self.app)
-            self.app.fi2tc_entries.append(dlg.data)
-            self.refresh()
-        def edit_row(self):
-            sel = self.tree.focus()
-            if not sel:
-                return
-            idx = self.tree.index(sel)
-            data = self.app.fi2tc_entries[idx]
-            dlg = self.RowDialog(self, self.app, data)
-            self.refresh()
-        def del_row(self):
-            sel = self.tree.selection()
-            for iid in sel:
-                idx = self.tree.index(iid)
-                if idx < len(self.app.fi2tc_entries):
-                    del self.app.fi2tc_entries[idx]
-            self.refresh()
-        def export_csv(self):
-            path = filedialog.asksaveasfilename(defaultextension=".csv", filetypes=[("CSV","*.csv")])
-            if not path:
-                return
-            with open(path, "w", newline="") as f:
-                w = csv.writer(f)
-                w.writerow(self.COLS)
-                for r in self.app.fi2tc_entries:
-                    w.writerow([r.get(k, "") for k in self.COLS])
-            messagebox.showinfo("Export", "FI2TC exported")
-
-    class HazopWindow(tk.Toplevel):
-        def __init__(self, app):
-            super().__init__(app.root)
-            self.app = app
-            self.title("HAZOP Analysis")
-            self.geometry("600x400")
-
-            columns = ("function", "malfunction", "type", "safety", "rationale")
-            self.tree = ttk.Treeview(self, columns=columns, show="headings")
-            for col in columns:
-                self.tree.heading(col, text=col.capitalize())
-                if col in ("rationale", "hazard"):
-                    width = 200
-                else:
-                    width = 120
-                self.tree.column(col, width=width)
-            self.tree.pack(fill=tk.BOTH, expand=True)
-
-            btn = ttk.Frame(self)
-            btn.pack(fill=tk.X)
-            ttk.Button(btn, text="Add", command=self.add_row).pack(side=tk.LEFT, padx=2, pady=2)
-            ttk.Button(btn, text="Edit", command=self.edit_row).pack(side=tk.LEFT, padx=2, pady=2)
-            ttk.Button(btn, text="Delete", command=self.del_row).pack(side=tk.LEFT, padx=2, pady=2)
-
-            self.refresh()
-
-        def refresh(self):
-            self.tree.delete(*self.tree.get_children())
-            for row in self.app.hazop_entries:
-                vals = [
-                    row.function,
-                    row.malfunction,
-                    row.mtype,
-                    row.scenario,
-                    row.conditions,
-                    row.hazard,
-                    "Yes" if row.safety else "No",
-                    row.rationale,
-                    "Yes" if row.covered else "No",
-                    row.covered_by,
-                ]
-                self.tree.insert("", "end", values=vals)
-
-        class RowDialog(simpledialog.Dialog):
-            def __init__(self, parent, row=None):
-                self.row = row or HazopEntry(
-                    "",
-                    "",
-                    "No/Not",
-                    "",
-                    "",
-                    "",
-                    False,
-                    "",
-                    False,
-                    "",
-                )
-                super().__init__(parent, title="Edit HAZOP Row")
-
-            def body(self, master):
-                ttk.Label(master, text="Function").grid(row=0, column=0, sticky="e", padx=5, pady=5)
-                self.func = tk.StringVar(value=self.row.function)
-                ttk.Entry(master, textvariable=self.func).grid(row=0, column=1, padx=5, pady=5)
-
-                ttk.Label(master, text="Malfunction").grid(row=1, column=0, sticky="e", padx=5, pady=5)
-                self.mal = tk.StringVar(value=self.row.malfunction)
-                ttk.Entry(master, textvariable=self.mal).grid(row=1, column=1, padx=5, pady=5)
-
-                ttk.Label(master, text="Type").grid(row=2, column=0, sticky="e", padx=5, pady=5)
-                self.typ = tk.StringVar(value=self.row.mtype)
-                ttk.Combobox(
-                    master,
-                    textvariable=self.typ,
-                    values=["No/Not", "Unintended", "Excessive", "Insufficient", "Reverse"],
-                    state="readonly",
-                ).grid(row=2, column=1, padx=5, pady=5)
-
-                ttk.Label(master, text="Scenario").grid(row=3, column=0, sticky="e", padx=5, pady=5)
-                scenarios = []
-                for lib in self.app.scenario_libraries:
-                    scenarios.extend(lib.get("scenarios", []))
-                self.scen = tk.StringVar(value=self.row.scenario)
-                ttk.Combobox(master, textvariable=self.scen, values=scenarios, state="readonly").grid(row=3, column=1, padx=5, pady=5)
-
-                ttk.Label(master, text="Driving Conditions").grid(row=4, column=0, sticky="e", padx=5, pady=5)
-                self.cond = tk.StringVar(value=self.row.conditions)
-                ttk.Entry(master, textvariable=self.cond).grid(row=4, column=1, padx=5, pady=5)
-
-                ttk.Label(master, text="Hazard").grid(row=5, column=0, sticky="ne", padx=5, pady=5)
-                self.haz = tk.Text(master, width=30, height=3)
-                self.haz.insert("1.0", self.row.hazard)
-                self.haz.grid(row=5, column=1, padx=5, pady=5)
-
-                ttk.Label(master, text="Safety Relevant").grid(row=6, column=0, sticky="e", padx=5, pady=5)
-                self.safety = tk.StringVar(value="Yes" if self.row.safety else "No")
-                ttk.Combobox(master, textvariable=self.safety, values=["Yes", "No"], state="readonly").grid(row=6, column=1, padx=5, pady=5)
-
-                ttk.Label(master, text="Rationale").grid(row=7, column=0, sticky="ne", padx=5, pady=5)
-                self.rat = tk.Text(master, width=30, height=3)
-                self.rat.insert("1.0", self.row.rationale)
-                self.rat.grid(row=7, column=1, padx=5, pady=5)
-
-                ttk.Label(master, text="Covered").grid(row=8, column=0, sticky="e", padx=5, pady=5)
-                self.cov = tk.StringVar(value="Yes" if self.row.covered else "No")
-                ttk.Combobox(master, textvariable=self.cov, values=["Yes", "No"], state="readonly").grid(row=8, column=1, padx=5, pady=5)
-
-                ttk.Label(master, text="Covered By").grid(row=9, column=0, sticky="e", padx=5, pady=5)
-                malfs = [e.malfunction for e in self.app.hazop_entries]
-                self.cov_by = tk.StringVar(value=self.row.covered_by)
-                ttk.Combobox(master, textvariable=self.cov_by, values=malfs, state="readonly").grid(row=9, column=1, padx=5, pady=5)
-
-            def apply(self):
-                self.row.function = self.func.get()
-                self.row.malfunction = self.mal.get()
-                self.row.mtype = self.typ.get()
-                self.row.scenario = self.scen.get()
-                self.row.conditions = self.cond.get()
-                self.row.hazard = self.haz.get("1.0", "end-1c")
-                self.row.safety = self.safety.get() == "Yes"
-                self.row.rationale = self.rat.get("1.0", "end-1c")
-                self.row.covered = self.cov.get() == "Yes"
-                self.row.covered_by = self.cov_by.get()
-
-        def add_row(self):
-            dlg = self.RowDialog(self)
-            if dlg.row.function:
-                self.app.hazop_entries.append(dlg.row)
-                self.refresh()
-
-        def edit_row(self):
-            sel = self.tree.focus()
-            if not sel:
-                return
-            idx = self.tree.index(sel)
-            row = self.app.hazop_entries[idx]
-            dlg = self.RowDialog(self, row)
-            self.refresh()
-
-        def del_row(self):
-            sel = self.tree.selection()
-            for iid in sel:
-                idx = self.tree.index(iid)
-                if idx < len(self.app.hazop_entries):
-                    del self.app.hazop_entries[idx]
-            self.refresh()
-
-        def load_analysis(self):
-            if not self.app.reliability_analyses:
-                messagebox.showwarning("Load", "No saved analyses")
-                return
-            win = tk.Toplevel(self)
-            win.title("Select Analysis")
-            lb = tk.Listbox(win, height=8, width=40)
-            lb.pack(side=tk.LEFT, fill=tk.BOTH, expand=True)
-            for ra in self.app.reliability_analyses:
-                lb.insert(tk.END, ra.name)
-            def do_load():
-                sel = lb.curselection()
-                if not sel:
-                    return
-                ra = self.app.reliability_analyses[sel[0]]
-                self.standard_var.set(ra.standard)
-                self.profile_var.set(ra.profile)
-                self.components = copy.deepcopy(ra.components)
-                self.app.reliability_total_fit = ra.total_fit
-                self.app.spfm = ra.spfm
-                self.app.lpfm = ra.lpfm
-                self.app.reliability_dc = ra.dc
-                win.destroy()
-                self.refresh_tree()
-                self.formula_label.config(
-                    text=f"Total FIT: {ra.total_fit:.2f}  DC: {ra.dc:.2f}  SPFM: {ra.spfm:.2f}  LPFM: {ra.lpfm:.2f}"
-                )
-            ttk.Button(win, text="Load", command=do_load).pack(side=tk.RIGHT, padx=5, pady=5)
-
-        def save_analysis(self):
-            if not self.components:
-                messagebox.showwarning("Save", "No components defined")
-                return
-            name = simpledialog.askstring("Save Analysis", "Enter analysis name:")
-            if not name:
-                return
-            ra = ReliabilityAnalysis(
-                name,
-                self.standard_var.get(),
-                self.profile_var.get(),
-                copy.deepcopy(self.components),
-                self.app.reliability_total_fit,
-                self.app.spfm,
-                self.app.lpfm,
-                self.app.reliability_dc,
-            )
-            self.app.reliability_analyses.append(ra)
-            messagebox.showinfo("Save", "Analysis saved")
-
-
-    class HaraWindow(tk.Toplevel):
-        COLS = [
-            "malfunction","severity","sev_rationale","controllability",
-            "cont_rationale","exposure","exp_rationale","asil","safety_goal"
-        ]
-
-        def __init__(self, app):
-            super().__init__(app.root)
-            self.app = app
-            self.title("HARA Analysis")
-            self.tree = ttk.Treeview(self, columns=self.COLS, show="headings")
-            for c in self.COLS:
-                self.tree.heading(c, text=c.replace("_"," ").title())
-                self.tree.column(c, width=120)
-            self.tree.pack(fill=tk.BOTH, expand=True)
-            btn = ttk.Frame(self)
-            btn.pack()
-            ttk.Button(btn, text="Add", command=self.add_row).pack(side=tk.LEFT, padx=2, pady=2)
-            ttk.Button(btn, text="Edit", command=self.edit_row).pack(side=tk.LEFT, padx=2, pady=2)
-            ttk.Button(btn, text="Delete", command=self.del_row).pack(side=tk.LEFT, padx=2, pady=2)
-            self.refresh()
-
-        def refresh(self):
-            self.tree.delete(*self.tree.get_children())
-            for row in self.app.hara_entries:
-                vals = [
-                    row.malfunction, row.severity, row.sev_rationale,
-                    row.controllability, row.cont_rationale,
-                    row.exposure, row.exp_rationale,
-                    row.asil, row.safety_goal
-                ]
-                self.tree.insert("", "end", values=vals)
-            self.app.sync_hara_to_safety_goals()
-
-        class RowDialog(simpledialog.Dialog):
-            def __init__(self, parent, app, row=None):
-                self.app = app
-                self.row = row or HaraEntry("",1,"",1,"",1,"","QM","")
-                super().__init__(parent, title="Edit HARA Row")
-
-            def body(self, master):
-                malfs = [e.malfunction for e in self.app.hazop_entries if e.safety]
-                goals = [te.safety_goal_description or (te.user_name or f"SG {te.unique_id}") for te in self.app.top_events]
-                ttk.Label(master, text="Malfunction").grid(row=0,column=0,sticky="e")
-                self.mal_var = tk.StringVar(value=self.row.malfunction)
-                ttk.Combobox(master, textvariable=self.mal_var, values=malfs, state="readonly").grid(row=0,column=1)
-                ttk.Label(master, text="Severity").grid(row=1,column=0,sticky="e")
-                self.sev_var = tk.StringVar(value=str(self.row.severity))
-                sev_cb = ttk.Combobox(master, textvariable=self.sev_var, values=["1","2","3"], state="readonly")
-                sev_cb.grid(row=1,column=1)
-                ttk.Label(master, text="Severity Rationale").grid(row=2,column=0,sticky="e")
-                self.sev_rat = tk.Entry(master)
-                self.sev_rat.insert(0, self.row.sev_rationale)
-                self.sev_rat.grid(row=2,column=1)
-                ttk.Label(master, text="Controllability").grid(row=3,column=0,sticky="e")
-                self.cont_var = tk.StringVar(value=str(self.row.controllability))
-                cont_cb = ttk.Combobox(master, textvariable=self.cont_var, values=["1","2","3"], state="readonly")
-                cont_cb.grid(row=3,column=1)
-                ttk.Label(master, text="Controllability Rationale").grid(row=4,column=0,sticky="e")
-                self.cont_rat = tk.Entry(master)
-                self.cont_rat.insert(0, self.row.cont_rationale)
-                self.cont_rat.grid(row=4,column=1)
-                ttk.Label(master, text="Exposure").grid(row=5,column=0,sticky="e")
-                self.exp_var = tk.StringVar(value=str(self.row.exposure))
-                exp_cb = ttk.Combobox(master, textvariable=self.exp_var, values=["1","2","3","4"], state="readonly")
-                exp_cb.grid(row=5,column=1)
-                ttk.Label(master, text="Exposure Rationale").grid(row=6,column=0,sticky="e")
-                self.exp_rat = tk.Entry(master)
-                self.exp_rat.insert(0, self.row.exp_rationale)
-                self.exp_rat.grid(row=6,column=1)
-                ttk.Label(master, text="ASIL").grid(row=7,column=0,sticky="e")
-                self.asil_var = tk.StringVar(value=self.row.asil)
-                asil_lbl = ttk.Label(master, textvariable=self.asil_var)
-                asil_lbl.grid(row=7,column=1)
-                ttk.Label(master, text="Safety Goal").grid(row=8,column=0,sticky="e")
-                self.sg_var = tk.StringVar(value=self.row.safety_goal)
-                ttk.Combobox(master, textvariable=self.sg_var, values=goals, state="readonly").grid(row=8,column=1)
-
-                def recalc(_=None):
-                    try:
-                        s = int(self.sev_var.get())
-                        c = int(self.cont_var.get())
-                        e = int(self.exp_var.get())
-                    except ValueError:
-                        self.asil_var.set("QM")
-                        return
-                    self.asil_var.set(calc_asil(s,c,e))
-
-                sev_cb.bind("<<ComboboxSelected>>", recalc)
-                cont_cb.bind("<<ComboboxSelected>>", recalc)
-                exp_cb.bind("<<ComboboxSelected>>", recalc)
-                recalc()
-
-            def apply(self):
-                self.row.malfunction = self.mal_var.get()
-                self.row.severity = int(self.sev_var.get())
-                self.row.sev_rationale = self.sev_rat.get()
-                self.row.controllability = int(self.cont_var.get())
-                self.row.cont_rationale = self.cont_rat.get()
-                self.row.exposure = int(self.exp_var.get())
-                self.row.exp_rationale = self.exp_rat.get()
-                self.row.asil = self.asil_var.get()
-                self.row.safety_goal = self.sg_var.get()
-
-        def add_row(self):
-            dlg = self.RowDialog(self, self.app)
-            self.app.hara_entries.append(dlg.row)
-            self.refresh()
-
-        def edit_row(self):
-            sel = self.tree.focus()
-            if not sel:
-                return
-            idx = self.tree.index(sel)
-            dlg = self.RowDialog(self, self.app, self.app.hara_entries[idx])
-            self.refresh()
-
-        def del_row(self):
-            sel = self.tree.selection()
-            for iid in sel:
-                idx = self.tree.index(iid)
-                if idx < len(self.app.hara_entries):
-                    del self.app.hara_entries[idx]
-            self.refresh()
-
-    class TC2FIWindow(tk.Toplevel):
-        COLS = [
-            "id",
-            "hazard_use_case",
-            "occurrence",
-            "system_function",
-            "elements_impacted",
-            "interfaces",
-            "functional_insufficiency",
-            "vehicle_effect",
-            "severity",
-            "design_measures",
-            "verification",
-            "measure_effectiveness",
-            "scene",
-            "scenario",
-            "driver_behavior",
-            "triggering_condition",
-            "tc_effect",
-            "mitigation",
-            "acceptance",
-        ]
-        def __init__(self, app):
-            super().__init__(app.root)
-            self.app = app
-            self.title("TC2FI Analysis")
-            self.tree = ttk.Treeview(self, columns=self.COLS, show="headings")
-            for c in self.COLS:
-                self.tree.heading(c, text=c.replace("_"," ").title())
-                self.tree.column(c, width=120)
-            self.tree.pack(fill=tk.BOTH, expand=True)
-            btn = ttk.Frame(self)
-            btn.pack()
-            ttk.Button(btn, text="Add", command=self.add_row).pack(side=tk.LEFT, padx=2, pady=2)
-            ttk.Button(btn, text="Edit", command=self.edit_row).pack(side=tk.LEFT, padx=2, pady=2)
-            ttk.Button(btn, text="Delete", command=self.del_row).pack(side=tk.LEFT, padx=2, pady=2)
-            ttk.Button(btn, text="Export CSV", command=self.export_csv).pack(side=tk.LEFT, padx=2, pady=2)
-            self.refresh()
-        def refresh(self):
-            self.tree.delete(*self.tree.get_children())
-            for row in self.app.tc2fi_entries:
-                self.tree.insert("", "end", values=[row.get(k, "") for k in self.COLS])
-        class RowDialog(simpledialog.Dialog):
-            def __init__(self, parent, app, data=None):
-                self.app = app
-                self.parent_win = parent
-                default = {k: "" for k in parent.COLS}
-                self.data = data or default
-                super().__init__(parent, title="Edit Row")
-            def body(self, master):
-                tc_names = [n.user_name or f"TC {n.unique_id}" for n in self.app.get_all_triggering_conditions()]
-                fi_names = [n.user_name or f"FI {n.unique_id}" for n in self.app.get_all_functional_insufficiencies()]
-                self.widgets = {}
-                r = 0
-                for col in self.parent_win.COLS:
-                    ttk.Label(master, text=col.replace("_"," ").title()).grid(row=r, column=0, sticky="e")
-                    if col == "functional_insufficiency":
-                        var = tk.StringVar(value=self.data.get(col, ""))
-                        cb = ttk.Combobox(master, textvariable=var, values=fi_names, state="readonly")
-                        cb.grid(row=r, column=1)
-                        self.widgets[col] = var
-                    elif col == "triggering_condition":
-                        var = tk.StringVar(value=self.data.get(col, ""))
-                        cb = ttk.Combobox(master, textvariable=var, values=tc_names, state="readonly")
-                        cb.grid(row=r, column=1)
-                        self.widgets[col] = var
-                    else:
-                        ent = tk.Entry(master)
-                        ent.insert(0, self.data.get(col, ""))
-                        ent.grid(row=r, column=1)
-                        self.widgets[col] = ent
-                    r += 1
-            def apply(self):
-                for col, widget in self.widgets.items():
-                    if isinstance(widget, tk.Entry):
-                        self.data[col] = widget.get()
-                    else:
-                        self.data[col] = widget.get()
-        def add_row(self):
-            dlg = self.RowDialog(self, self.app)
-            self.app.tc2fi_entries.append(dlg.data)
-            self.refresh()
-        def edit_row(self):
-            sel = self.tree.focus()
-            if not sel:
-                return
-            idx = self.tree.index(sel)
-            data = self.app.tc2fi_entries[idx]
-            dlg = self.RowDialog(self, self.app, data)
-            self.refresh()
-        def del_row(self):
-            sel = self.tree.selection()
-            for iid in sel:
-                idx = self.tree.index(iid)
-                if idx < len(self.app.tc2fi_entries):
-                    del self.app.tc2fi_entries[idx]
-            self.refresh()
-        def export_csv(self):
-            path = filedialog.asksaveasfilename(defaultextension=".csv", filetypes=[("CSV","*.csv")])
-            if not path:
-                return
-            with open(path, "w", newline="") as f:
-                w = csv.writer(f)
-                w.writerow(self.COLS)
-                for r in self.app.tc2fi_entries:
-                    w.writerow([r.get(k, "") for k in self.COLS])
-            messagebox.showinfo("Export", "TC2FI exported")
-
-
-    class TC2FIWindow(tk.Toplevel):
-        COLS = [
-            "id","system_function","allocation","interfaces","triggering_condition",
-            "scene","scenario","driver_behavior","occurrence","vehicle_effect",
-            "severity","design_measures","verification","measure_effectiveness",
-            "functional_insufficiency","worst_case","fi_effect","mitigation","acceptance"
-        ]
-        def __init__(self, app):
-            super().__init__(app.root)
-            self.app = app
-            self.title("TC2FI Analysis")
-            self.tree = ttk.Treeview(self, columns=self.COLS, show="headings")
-            for c in self.COLS:
-                self.tree.heading(c, text=c.replace("_"," ").title())
-                self.tree.column(c, width=120)
-            self.tree.pack(fill=tk.BOTH, expand=True)
-            btn = ttk.Frame(self)
-            btn.pack()
-            ttk.Button(btn, text="Add", command=self.add_row).pack(side=tk.LEFT, padx=2, pady=2)
-            ttk.Button(btn, text="Edit", command=self.edit_row).pack(side=tk.LEFT, padx=2, pady=2)
-            ttk.Button(btn, text="Delete", command=self.del_row).pack(side=tk.LEFT, padx=2, pady=2)
-            ttk.Button(btn, text="Export CSV", command=self.export_csv).pack(side=tk.LEFT, padx=2, pady=2)
-            self.refresh()
-        def refresh(self):
-            self.tree.delete(*self.tree.get_children())
-            for row in self.app.tc2fi_entries:
-                self.tree.insert("", "end", values=[row.get(k, "") for k in self.COLS])
-        class RowDialog(simpledialog.Dialog):
-            def __init__(self, parent, app, data=None):
-                self.app = app
-                default = {k: "" for k in TC2FIWindow.COLS}
-                self.data = data or default
-                super().__init__(parent, title="Edit Row")
-            def body(self, master):
-                tc_names = [n.user_name or f"TC {n.unique_id}" for n in self.app.get_all_triggering_conditions()]
-                fi_names = [n.user_name or f"FI {n.unique_id}" for n in self.app.get_all_functional_insufficiencies()]
-                self.widgets = {}
-                r = 0
-                for col in TC2FIWindow.COLS:
-                    ttk.Label(master, text=col.replace("_"," ").title()).grid(row=r, column=0, sticky="e")
-                    if col == "functional_insufficiency":
-                        var = tk.StringVar(value=self.data.get(col, ""))
-                        cb = ttk.Combobox(master, textvariable=var, values=fi_names, state="readonly")
-                        cb.grid(row=r, column=1)
-                        self.widgets[col] = var
-                    elif col == "triggering_condition":
-                        var = tk.StringVar(value=self.data.get(col, ""))
-                        cb = ttk.Combobox(master, textvariable=var, values=tc_names, state="readonly")
-                        cb.grid(row=r, column=1)
-                        self.widgets[col] = var
-                    else:
-                        ent = tk.Entry(master)
-                        ent.insert(0, self.data.get(col, ""))
-                        ent.grid(row=r, column=1)
-                        self.widgets[col] = ent
-                    r += 1
-            def apply(self):
-                for col, widget in self.widgets.items():
-                    if isinstance(widget, tk.Entry):
-                        self.data[col] = widget.get()
-                    else:
-                        self.data[col] = widget.get()
-        def add_row(self):
-            dlg = self.RowDialog(self, self.app)
-            self.app.tc2fi_entries.append(dlg.data)
-            self.refresh()
-        def edit_row(self):
-            sel = self.tree.focus()
-            if not sel:
-                return
-            idx = self.tree.index(sel)
-            data = self.app.tc2fi_entries[idx]
-            dlg = self.RowDialog(self, self.app, data)
-            self.refresh()
-        def del_row(self):
-            sel = self.tree.selection()
-            for iid in sel:
-                idx = self.tree.index(iid)
-                if idx < len(self.app.tc2fi_entries):
-                    del self.app.tc2fi_entries[idx]
-            self.refresh()
-        def export_csv(self):
-            path = filedialog.asksaveasfilename(defaultextension=".csv", filetypes=[("CSV","*.csv")])
-            if not path:
-                return
-            with open(path, "w", newline="") as f:
-                w = csv.writer(f)
-                w.writerow(self.COLS)
-                for r in self.app.tc2fi_entries:
-                    w.writerow([r.get(k, "") for k in self.COLS])
-            messagebox.showinfo("Export", "TC2FI exported")
-
-        def __init__(self, app):
-            super().__init__(app.root)
-            self.app = app
-            self.title("FI2TC Analysis")
-            cols = ("fi","tc","scenario","hazard","mitigation","asil")
-            self.tree = ttk.Treeview(self, columns=cols, show="headings")
-            for c in cols:
-                self.tree.heading(c, text=c.capitalize())
-                self.tree.column(c, width=100)
-            self.tree.pack(fill=tk.BOTH, expand=True)
-            btn = ttk.Frame(self)
-            btn.pack()
-            ttk.Button(btn, text="Add", command=self.add_row).pack(side=tk.LEFT, padx=2, pady=2)
-            ttk.Button(btn, text="Edit", command=self.edit_row).pack(side=tk.LEFT, padx=2, pady=2)
-            ttk.Button(btn, text="Delete", command=self.del_row).pack(side=tk.LEFT, padx=2, pady=2)
-            ttk.Button(btn, text="Export CSV", command=self.export_csv).pack(side=tk.LEFT, padx=2, pady=2)
-            self.refresh()
-
-        def refresh(self):
-            self.tree.delete(*self.tree.get_children())
-            for row in self.app.fi2tc_entries:
-                self.tree.insert("", "end", values=(row.get("fi"), row.get("tc"), row.get("scenario"), row.get("hazard"), row.get("mitigation"), row.get("asil")))
-
-        class RowDialog(simpledialog.Dialog):
-            def __init__(self, parent, app, data=None):
-                self.app = app
-                self.data = data or {"fi":"","tc":"","scenario":"","hazard":"","mitigation":"","asil":""}
-                super().__init__(parent, title="Edit Row")
-            def body(self, master):
-                fi_names = [n.user_name or f"FI {n.unique_id}" for n in self.app.get_all_functional_insufficiencies()]
-                tc_names = [n.user_name or f"TC {n.unique_id}" for n in self.app.get_all_triggering_conditions()]
-                self.fi_var = tk.StringVar(value=self.data.get("fi"))
-                ttk.Combobox(master, textvariable=self.fi_var, values=fi_names, state="readonly").grid(row=0,column=1)
-                ttk.Label(master, text="Functional Insufficiency").grid(row=0,column=0,sticky="e")
-                self.tc_var = tk.StringVar(value=self.data.get("tc"))
-                ttk.Combobox(master, textvariable=self.tc_var, values=tc_names, state="readonly").grid(row=1,column=1)
-                ttk.Label(master, text="Triggering Condition").grid(row=1,column=0,sticky="e")
-                ttk.Label(master, text="Scenario").grid(row=2,column=0,sticky="e")
-                self.sc_var = tk.Entry(master)
-                self.sc_var.insert(0, self.data.get("scenario"))
-                self.sc_var.grid(row=2,column=1)
-                ttk.Label(master, text="Hazard").grid(row=3,column=0,sticky="e")
-                self.haz_var = tk.Entry(master)
-                self.haz_var.insert(0, self.data.get("hazard"))
-                self.haz_var.grid(row=3,column=1)
-                ttk.Label(master, text="Mitigation").grid(row=4,column=0,sticky="e")
-                self.mit_var = tk.Entry(master)
-                self.mit_var.insert(0, self.data.get("mitigation"))
-                self.mit_var.grid(row=4,column=1)
-                ttk.Label(master, text="ASIL").grid(row=5,column=0,sticky="e")
-                self.asil_var = tk.Entry(master)
-                self.asil_var.insert(0, self.data.get("asil"))
-                self.asil_var.grid(row=5,column=1)
-            def apply(self):
-                self.data["fi"] = self.fi_var.get()
-                self.data["tc"] = self.tc_var.get()
-                self.data["scenario"] = self.sc_var.get()
-                self.data["hazard"] = self.haz_var.get()
-                self.data["mitigation"] = self.mit_var.get()
-                self.data["asil"] = self.asil_var.get()
-
-        def add_row(self):
-            dlg = self.RowDialog(self, self.app)
-            self.app.fi2tc_entries.append(dlg.data)
-            self.refresh()
-        def edit_row(self):
-            sel = self.tree.focus()
-            if not sel: return
-            idx = self.tree.index(sel)
-            data = self.app.fi2tc_entries[idx]
-            dlg = self.RowDialog(self, self.app, data)
-            self.refresh()
-        def del_row(self):
-            sel = self.tree.selection()
-            for iid in sel:
-                idx = self.tree.index(iid)
-                if idx < len(self.app.fi2tc_entries):
-                    del self.app.fi2tc_entries[idx]
-            self.refresh()
-        def export_csv(self):
-            path = filedialog.asksaveasfilename(defaultextension=".csv", filetypes=[("CSV","*.csv")])
-            if not path: return
-            with open(path,"w",newline="") as f:
-                w=csv.writer(f)
-                w.writerow(["Functional Insufficiency","Triggering Condition","Scenario","Hazard","Mitigation","ASIL"])
-                for r in self.app.fi2tc_entries:
-                    w.writerow([r.get("fi"),r.get("tc"),r.get("scenario"),r.get("hazard"),r.get("mitigation"),r.get("asil")])
-            messagebox.showinfo("Export","FI2TC exported")
-
-    class TC2FIWindow(tk.Toplevel):
-        def __init__(self, app):
-            super().__init__(app.root)
-            self.app = app
-            self.title("TC2FI Analysis")
-            cols = ("tc","fi","scenario","hazard","mitigation","asil")
-            self.tree = ttk.Treeview(self, columns=cols, show="headings")
-            for c in cols:
-                self.tree.heading(c, text=c.capitalize())
-                self.tree.column(c, width=100)
-            self.tree.pack(fill=tk.BOTH, expand=True)
-            btn = ttk.Frame(self)
-            btn.pack()
-            ttk.Button(btn, text="Add", command=self.add_row).pack(side=tk.LEFT, padx=2, pady=2)
-            ttk.Button(btn, text="Edit", command=self.edit_row).pack(side=tk.LEFT, padx=2, pady=2)
-            ttk.Button(btn, text="Delete", command=self.del_row).pack(side=tk.LEFT, padx=2, pady=2)
-            ttk.Button(btn, text="Export CSV", command=self.export_csv).pack(side=tk.LEFT, padx=2, pady=2)
-            self.refresh()
-        def refresh(self):
-            self.tree.delete(*self.tree.get_children())
-            for row in self.app.tc2fi_entries:
-                self.tree.insert("", "end", values=(row.get("tc"), row.get("fi"), row.get("scenario"), row.get("hazard"), row.get("mitigation"), row.get("asil")))
-        class RowDialog(simpledialog.Dialog):
-            def __init__(self, parent, app, data=None):
-                self.app=app
-                self.data=data or {"tc":"","fi":"","scenario":"","hazard":"","mitigation":"","asil":""}
-                super().__init__(parent, title="Edit Row")
-            def body(self, master):
-                tc_names=[n.user_name or f"TC {n.unique_id}" for n in self.app.get_all_triggering_conditions()]
-                fi_names=[n.user_name or f"FI {n.unique_id}" for n in self.app.get_all_functional_insufficiencies()]
-                self.tc_var=tk.StringVar(value=self.data.get("tc"))
-                ttk.Combobox(master,textvariable=self.tc_var,values=tc_names,state="readonly").grid(row=0,column=1)
-                ttk.Label(master,text="Triggering Condition").grid(row=0,column=0,sticky="e")
-                self.fi_var=tk.StringVar(value=self.data.get("fi"))
-                ttk.Combobox(master,textvariable=self.fi_var,values=fi_names,state="readonly").grid(row=1,column=1)
-                ttk.Label(master,text="Functional Insufficiency").grid(row=1,column=0,sticky="e")
-                ttk.Label(master,text="Scenario").grid(row=2,column=0,sticky="e")
-                self.sc_var=tk.Entry(master);self.sc_var.insert(0,self.data.get("scenario"));self.sc_var.grid(row=2,column=1)
-                ttk.Label(master,text="Hazard").grid(row=3,column=0,sticky="e")
-                self.haz_var=tk.Entry(master);self.haz_var.insert(0,self.data.get("hazard"));self.haz_var.grid(row=3,column=1)
-                ttk.Label(master,text="Mitigation").grid(row=4,column=0,sticky="e")
-                self.mit_var=tk.Entry(master);self.mit_var.insert(0,self.data.get("mitigation"));self.mit_var.grid(row=4,column=1)
-                ttk.Label(master,text="ASIL").grid(row=5,column=0,sticky="e")
-                self.asil_var=tk.Entry(master);self.asil_var.insert(0,self.data.get("asil"));self.asil_var.grid(row=5,column=1)
-            def apply(self):
-                self.data["tc"]=self.tc_var.get();self.data["fi"]=self.fi_var.get();self.data["scenario"]=self.sc_var.get();self.data["hazard"]=self.haz_var.get();self.data["mitigation"]=self.mit_var.get();self.data["asil"]=self.asil_var.get()
-        def add_row(self):
-            dlg=self.RowDialog(self,self.app)
-            self.app.tc2fi_entries.append(dlg.data);self.refresh()
-        def edit_row(self):
-            sel=self.tree.focus()
-            if not sel: return
-            idx=self.tree.index(sel);data=self.app.tc2fi_entries[idx];dlg=self.RowDialog(self,self.app,data);self.refresh()
-        def del_row(self):
-            sel=self.tree.selection()
-            for iid in sel:
-                idx=self.tree.index(iid)
-                if idx < len(self.app.tc2fi_entries):
-                    del self.app.tc2fi_entries[idx]
-            self.refresh()
-        def export_csv(self):
-            path=filedialog.asksaveasfilename(defaultextension=".csv", filetypes=[("CSV","*.csv")])
-            if not path: return
-            with open(path,"w",newline="") as f:
-                w=csv.writer(f);w.writerow(["Triggering Condition","Functional Insufficiency","Scenario","Hazard","Mitigation","ASIL"]);
-                for r in self.app.tc2fi_entries:
-                    w.writerow([r.get("tc"),r.get("fi"),r.get("scenario"),r.get("hazard"),r.get("mitigation"),r.get("asil")])
-            messagebox.showinfo("Export","TC2FI exported")
-
-        def __init__(self, app):
-            super().__init__(app.root)
-            self.app = app
-            self.title("FI2TC Analysis")
-            cols = ("fi","tc","scenario","hazard","mitigation","asil")
-            self.tree = ttk.Treeview(self, columns=cols, show="headings")
-            for c in cols:
-                self.tree.heading(c, text=c.capitalize())
-                self.tree.column(c, width=100)
-            self.tree.pack(fill=tk.BOTH, expand=True)
-            btn = ttk.Frame(self)
-            btn.pack()
-            ttk.Button(btn, text="Add", command=self.add_row).pack(side=tk.LEFT, padx=2, pady=2)
-            ttk.Button(btn, text="Edit", command=self.edit_row).pack(side=tk.LEFT, padx=2, pady=2)
-            ttk.Button(btn, text="Delete", command=self.del_row).pack(side=tk.LEFT, padx=2, pady=2)
-            ttk.Button(btn, text="Export CSV", command=self.export_csv).pack(side=tk.LEFT, padx=2, pady=2)
-            self.refresh()
-
-        def refresh(self):
-            self.tree.delete(*self.tree.get_children())
-            for row in self.app.fi2tc_entries:
-                self.tree.insert("", "end", values=(row.get("fi"), row.get("tc"), row.get("scenario"), row.get("hazard"), row.get("mitigation"), row.get("asil")))
-
-        class RowDialog(simpledialog.Dialog):
-            def __init__(self, parent, app, data=None):
-                self.app = app
-                self.data = data or {"fi":"","tc":"","scenario":"","hazard":"","mitigation":"","asil":""}
-                super().__init__(parent, title="Edit Row")
-            def body(self, master):
-                fi_names = [n.user_name or f"FI {n.unique_id}" for n in self.app.get_all_functional_insufficiencies()]
-                tc_names = [n.user_name or f"TC {n.unique_id}" for n in self.app.get_all_triggering_conditions()]
-                self.fi_var = tk.StringVar(value=self.data.get("fi"))
-                ttk.Combobox(master, textvariable=self.fi_var, values=fi_names, state="readonly").grid(row=0,column=1)
-                ttk.Label(master, text="Functional Insufficiency").grid(row=0,column=0,sticky="e")
-                self.tc_var = tk.StringVar(value=self.data.get("tc"))
-                ttk.Combobox(master, textvariable=self.tc_var, values=tc_names, state="readonly").grid(row=1,column=1)
-                ttk.Label(master, text="Triggering Condition").grid(row=1,column=0,sticky="e")
-                ttk.Label(master, text="Scenario").grid(row=2,column=0,sticky="e")
-                self.sc_var = tk.Entry(master)
-                self.sc_var.insert(0, self.data.get("scenario"))
-                self.sc_var.grid(row=2,column=1)
-                ttk.Label(master, text="Hazard").grid(row=3,column=0,sticky="e")
-                self.haz_var = tk.Entry(master)
-                self.haz_var.insert(0, self.data.get("hazard"))
-                self.haz_var.grid(row=3,column=1)
-                ttk.Label(master, text="Mitigation").grid(row=4,column=0,sticky="e")
-                self.mit_var = tk.Entry(master)
-                self.mit_var.insert(0, self.data.get("mitigation"))
-                self.mit_var.grid(row=4,column=1)
-                ttk.Label(master, text="ASIL").grid(row=5,column=0,sticky="e")
-                self.asil_var = tk.Entry(master)
-                self.asil_var.insert(0, self.data.get("asil"))
-                self.asil_var.grid(row=5,column=1)
-            def apply(self):
-                self.data["fi"] = self.fi_var.get()
-                self.data["tc"] = self.tc_var.get()
-                self.data["scenario"] = self.sc_var.get()
-                self.data["hazard"] = self.haz_var.get()
-                self.data["mitigation"] = self.mit_var.get()
-                self.data["asil"] = self.asil_var.get()
-
-        def add_row(self):
-            dlg = self.RowDialog(self, self.app)
-            self.app.fi2tc_entries.append(dlg.data)
-            self.refresh()
-        def edit_row(self):
-            sel = self.tree.focus()
-            if not sel: return
-            idx = self.tree.index(sel)
-            data = self.app.fi2tc_entries[idx]
-            dlg = self.RowDialog(self, self.app, data)
-            self.refresh()
-        def del_row(self):
-            sel = self.tree.selection()
-            for iid in sel:
-                idx = self.tree.index(iid)
-                if idx < len(self.app.fi2tc_entries):
-                    del self.app.fi2tc_entries[idx]
-            self.refresh()
-        def export_csv(self):
-            path = filedialog.asksaveasfilename(defaultextension=".csv", filetypes=[("CSV","*.csv")])
-            if not path: return
-            with open(path,"w",newline="") as f:
-                w=csv.writer(f)
-                w.writerow(["Functional Insufficiency","Triggering Condition","Scenario","Hazard","Mitigation","ASIL"])
-                for r in self.app.fi2tc_entries:
-                    w.writerow([r.get("fi"),r.get("tc"),r.get("scenario"),r.get("hazard"),r.get("mitigation"),r.get("asil")])
-            messagebox.showinfo("Export","FI2TC exported")
-
-    class TC2FIWindow(tk.Toplevel):
-        def __init__(self, app):
-            super().__init__(app.root)
-            self.app = app
-            self.title("TC2FI Analysis")
-            cols = ("tc","fi","scenario","hazard","mitigation","asil")
-            self.tree = ttk.Treeview(self, columns=cols, show="headings")
-            for c in cols:
-                self.tree.heading(c, text=c.capitalize())
-                self.tree.column(c, width=100)
-            self.tree.pack(fill=tk.BOTH, expand=True)
-            btn = ttk.Frame(self)
-            btn.pack()
-            ttk.Button(btn, text="Add", command=self.add_row).pack(side=tk.LEFT, padx=2, pady=2)
-            ttk.Button(btn, text="Edit", command=self.edit_row).pack(side=tk.LEFT, padx=2, pady=2)
-            ttk.Button(btn, text="Delete", command=self.del_row).pack(side=tk.LEFT, padx=2, pady=2)
-            ttk.Button(btn, text="Export CSV", command=self.export_csv).pack(side=tk.LEFT, padx=2, pady=2)
-            self.refresh()
-        def refresh(self):
-            self.tree.delete(*self.tree.get_children())
-            for row in self.app.tc2fi_entries:
-                self.tree.insert("", "end", values=(row.get("tc"), row.get("fi"), row.get("scenario"), row.get("hazard"), row.get("mitigation"), row.get("asil")))
-        class RowDialog(simpledialog.Dialog):
-            def __init__(self, parent, app, data=None):
-                self.app=app
-                self.data=data or {"tc":"","fi":"","scenario":"","hazard":"","mitigation":"","asil":""}
-                super().__init__(parent, title="Edit Row")
-            def body(self, master):
-                tc_names=[n.user_name or f"TC {n.unique_id}" for n in self.app.get_all_triggering_conditions()]
-                fi_names=[n.user_name or f"FI {n.unique_id}" for n in self.app.get_all_functional_insufficiencies()]
-                self.tc_var=tk.StringVar(value=self.data.get("tc"))
-                ttk.Combobox(master,textvariable=self.tc_var,values=tc_names,state="readonly").grid(row=0,column=1)
-                ttk.Label(master,text="Triggering Condition").grid(row=0,column=0,sticky="e")
-                self.fi_var=tk.StringVar(value=self.data.get("fi"))
-                ttk.Combobox(master,textvariable=self.fi_var,values=fi_names,state="readonly").grid(row=1,column=1)
-                ttk.Label(master,text="Functional Insufficiency").grid(row=1,column=0,sticky="e")
-                ttk.Label(master,text="Scenario").grid(row=2,column=0,sticky="e")
-                self.sc_var=tk.Entry(master);self.sc_var.insert(0,self.data.get("scenario"));self.sc_var.grid(row=2,column=1)
-                ttk.Label(master,text="Hazard").grid(row=3,column=0,sticky="e")
-                self.haz_var=tk.Entry(master);self.haz_var.insert(0,self.data.get("hazard"));self.haz_var.grid(row=3,column=1)
-                ttk.Label(master,text="Mitigation").grid(row=4,column=0,sticky="e")
-                self.mit_var=tk.Entry(master);self.mit_var.insert(0,self.data.get("mitigation"));self.mit_var.grid(row=4,column=1)
-                ttk.Label(master,text="ASIL").grid(row=5,column=0,sticky="e")
-                self.asil_var=tk.Entry(master);self.asil_var.insert(0,self.data.get("asil"));self.asil_var.grid(row=5,column=1)
-            def apply(self):
-                self.data["tc"]=self.tc_var.get();self.data["fi"]=self.fi_var.get();self.data["scenario"]=self.sc_var.get();self.data["hazard"]=self.haz_var.get();self.data["mitigation"]=self.mit_var.get();self.data["asil"]=self.asil_var.get()
-        def add_row(self):
-            dlg=self.RowDialog(self,self.app)
-            self.app.tc2fi_entries.append(dlg.data);self.refresh()
-        def edit_row(self):
-            sel=self.tree.focus()
-            if not sel: return
-            idx=self.tree.index(sel);data=self.app.tc2fi_entries[idx];dlg=self.RowDialog(self,self.app,data);self.refresh()
-        def del_row(self):
-            sel=self.tree.selection()
-            for iid in sel:
-                idx=self.tree.index(iid)
-                if idx < len(self.app.tc2fi_entries):
-                    del self.app.tc2fi_entries[idx]
-            self.refresh()
-        def export_csv(self):
-            path=filedialog.asksaveasfilename(defaultextension=".csv", filetypes=[("CSV","*.csv")])
-            if not path: return
-            with open(path,"w",newline="") as f:
-                w=csv.writer(f);w.writerow(["Triggering Condition","Functional Insufficiency","Scenario","Hazard","Mitigation","ASIL"]);
-                for r in self.app.tc2fi_entries:
-                    w.writerow([r.get("tc"),r.get("fi"),r.get("scenario"),r.get("hazard"),r.get("mitigation"),r.get("asil")])
-            messagebox.showinfo("Export","TC2FI exported")
-
     def copy_node(self):
         if self.selected_node and self.selected_node != self.root_node:
             self.clipboard_node = self.selected_node
